% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{active_connection}
\alias{active_connection}
\title{Active Connection}
\usage{
active_connection(con = NULL)
}
\arguments{
\item{con}{optional \code{\link{OpenEOClient}} to set, if omitted or NULL the currently active 
connection is returned}
}
\value{
\code{\link{OpenEOClient}}
}
\description{
The function gets or sets the currently active connection to an openEO service. Usually the
active connection is set when calling the \code{\link{connect}} function. Only the last 
connection is set to active.
An application for the active connection is the optional connection within all the functions
that interact with the openEO service and require a connection. If the connection is omitted 
in those function this function will be called in order to try to fetch a connection. If you 
want to operate on multiple services at once, you should use an explicit connection.
}
\seealso{
\code{\link{connect}}
}
