% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{api.version}
\alias{api.version}
\title{Returns the API version}
\usage{
api.version()
}
\value{
character describing the API version
}
\description{
This function returns information against which was developed in this R-client version.
}
