% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{Array}
\alias{Array}
\title{Array}
\value{
Object of \code{\link[R6:R6Class]{R6::R6Class()}} representing a single valued array.
}
\description{
Inheriting from \link{Argument} in order to represent an array of a single data type.
}
\section{Methods}{

\describe{
\item{\verb{$getMinItems}}{returns the minimum number of items}
\item{\verb{$getMaxItems}}{returns the maximum number of items}
\item{\verb{$setMinItems(value)}}{sets the minimum number of items}
\item{\verb{$setMaxItems(value)}}{sets the maximum number of items}
\item{\verb{$getItemSchema}}{returns the item schema of the items in the array}
\item{\verb{$setItemSchema(value)}}{sets the schema for the items in the array}
}
}

\section{Arguments}{

\describe{
\item{\code{value}}{either a number describing the minimum and maximum number of elements in an array or the
parsed JSON schema of a single item in the array}
}
}

\seealso{
\link{Array}, \link{Integer}, \link{EPSGCode}, \link{String}, \link{Number},
\link{Date}, \link{DataCube}, \link{ProcessGraphArgument},
\link{ProcessGraphParameter}, \link{OutputFormatOptions}, \link{GeoJson},
\link{Boolean}, \link{DateTime}, \link{Time}, \link{BoundingBox}, \link{Kernel},
\link{TemporalInterval}, \link{TemporalIntervals}, \link{CollectionId}, \link{OutputFormat},
\link{AnyOf}, \link{ProjDefinition}, \link{UdfCodeArgument}, \link{UdfRuntimeArgument} and
\link{UdfRuntimeVersionArgument}, \link{MetadataFilter}
}
