% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{as.bbox}
\alias{as.bbox}
\alias{as.bbox.bounding-box}
\alias{as.bbox.geojson}
\title{coerce to bbox}
\usage{
`as.bbox.bounding-box`(from)

as.bbox.geojson(from)
}
\arguments{
\item{from}{a \code{\link[=BoundingBox]{BoundingBox()}} argument object or a \code{\link[=GeoJson]{GeoJson()}} argument object}
}
\value{
a bbox object from \code{\link[sf:st_bbox]{sf::st_bbox()}}
}
\description{
A coercion function for extracting a 'bbox' object that can usualy be obtained by \code{\link[sf:st_bbox]{sf::st_bbox()}}. This coercion
function was created to easily extract the boúnding box from the openeos argument objects \code{\link[=BoundingBox]{BoundingBox()}} and \code{\link[=GeoJson]{GeoJson()}}.
}
