% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-functions.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame.JobList}
\alias{as.data.frame.ServiceList}
\alias{as.data.frame.BandList}
\alias{as.data.frame.CollectionList}
\alias{as.data.frame.VersionsList}
\alias{as.data.frame.FileFormatList}
\title{Coercions into data.frame objects}
\usage{
\method{as.data.frame}{JobList}(x, ...)

\method{as.data.frame}{ServiceList}(x, ...)

\method{as.data.frame}{BandList}(x, ...)

\method{as.data.frame}{CollectionList}(x, ...)

\method{as.data.frame}{VersionsList}(x, ...)

\method{as.data.frame}{FileFormatList}(x, ...)
}
\arguments{
\item{x}{the list object that will be coerced}

\item{...}{potentially additional parameters to pass on to internal functions like 'extract'}
}
\value{
a data.frame
}
\description{
The openeo package offers functions to transform list objects obtained from JSON
into data.frames. This is mostly applied in list_* functions.
}
\details{
The parameter 'extract' is used as an additional parameter to extract specific values of the output list
/ json. The value for the parameters is a vector of character like c('id','title')
}
