% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-functions.R
\name{as.Graph}
\alias{as.Graph}
\alias{as.Graph.ProcessNode}
\alias{as.Graph.function}
\alias{as.Graph.ProcessInfo}
\title{Coercion into Graph}
\usage{
as.Graph.ProcessNode(from)

as.Graph.function(from)

as.Graph.ProcessInfo(from)
}
\arguments{
\item{from}{the source from which to coerce (\code{ProcessNode}, \code{function} or \code{ProcessInfo})}
}
\value{
\code{\link{Graph}}
}
\description{
Creates a \code{Graph} object from a \code{\link{ProcessNode}}, \code{function} or \code{ProcessInfo} (Exchange object for 
predefined and stored user-defined processes).
}
\details{
Those pure \code{Graph} objects shall only be used internally. If you want to use this 
information in terms of interacting with the back-end directly via JSON please use 
\code{\link{as.Process}}. This function might be removed from the package function export in the future.
}
