% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-functions.R
\name{as.Process}
\alias{as.Process}
\alias{as.Process.ProcessInfo}
\title{Coerce into a Process}
\usage{
as.Process.ProcessInfo(from)
}
\arguments{
\item{from}{the source from which to coerce (\code{ProcessInfo})}
}
\value{
\code{\link{Process}}
}
\description{
This function converts objects into a process.
}
