% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate}
\usage{
authenticate(con, ...)
}
\arguments{
\item{con}{Connection object}
}
\value{
authenticated Connection
}
\description{
Authenticate
}
