% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band_arithmetic-functions.R
\name{band_arithmetics}
\alias{band_arithmetics}
\title{Performs band arithmetic on a data set}
\usage{
band_arithmetics(graph, data, formular)
}
\arguments{
\item{graph}{The graph into which the spectral reduce function will go}

\item{data}{A process that returns a data cube}

\item{formular}{A function that contains a mathematical expression}
}
\description{
The function evaluates a function with a mathematical expression and creates internally a
reduce function on the spectral dimension, by converting the function stated in 'formular'
into a subgraph for the callback.
}
\details{
Similar to the 'calc' function in the raster package, band arithmetic will be performed on an
data vector. The select bands are available via indexing in the R way (e.g. x[1] for the first
value). It is important that if the bands are prefiltered the order of the bands matters otherwise 
it is assumed that the bands are ordered as in the collection description.
The expression stated in the function shall be single line and should not contain
a return statement.
}
\note{
It might occur that a back-end does not comply to the process definitions of the core
process definition. This is when this function will not work properly. Alternatively you
have to manage the graph creation by your own and work with the given processes of the back-end.
}
\examples{
\dontrun{
...
data = graph$load_collection(graph$data$`COPERNICUS/S2`,
  spatial_extent = list(
     west=16.1,
     east=16.6,
     north=48.6,
     south= 47.2
   ),
   temporal_extent = list(
     "2018-01-01", "2018-02-01"
   ),
  bands=list("B08","B04","B02"))
 
evi = graph \%>\% band_arithmetics(data=data,function(x){
   2.5*((x[1]-x[2])/sum(1,x[1],6*x[2],-7.5*x[3]))
})

...
}
}
