% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback-functions.R
\name{callback}
\alias{callback}
\title{Creates a callback}
\usage{
callback(con, process, parameter = NULL, choice_index = NULL)
}
\arguments{
\item{con}{a connected \code{\link{OpenEOClient}}}

\item{process}{a \code{\link{Process}} or \code{\link{ProcessNode}} object of a back-end process}

\item{parameter}{optional name of a parameter of the process which requires a callback as value. If omitted then it returns only the names of parameter that require a callback}

\item{choice_index}{optional integer denoting the callback parameter selected in an \code{\link{AnyOf}}}
}
\value{
a \code{\link{Graph}} object with the callback parameters as 'data'
}
\description{
The callback function creates a callback graph for a parameter of a process if the parameter needs
a \code{\link{CallbackArgument}} as a value. If the parameter is omitted it shows possible parameter names which require
a callback.
}
