% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{CallbackArgument}
\alias{CallbackArgument}
\title{Callback}
\value{
Object of \code{\link{R6Class}} which represents a callback.
}
\description{
Inheriting from \code{\link{Argument}} in order to represent a callback. The callback operates on reduced data
of a data cube. For example reducing the time dimension results in a time series that has to be reduced into a
single value. The value of a callback is usually a \code{\link{Graph}} with \code{\link{CallbackValue}} as 
injected data. Hints from the openeo api documention:
\itemize{
  \item \url{https://open-eo.github.io/openeo-api/processes/#callbacks}
  \item \url{https://open-eo.github.io/openeo-api/v/0.4.2/processgraphs/#callbacks}
}
}
\section{Methods}{

\describe{
  \item{\code{$getCallbackParameters()}}{returns the available list \code{\link{CallbackValue}}}
  \item{\code{$setCallbackParameters(parameters)}}{assigns a list of \code{\link{CallbackValue}} to the callback}
}
}

\section{Arguments}{

\describe{
  \item{\code{parameters}}{the list \code{\link{CallbackValue}}}
}
}

\seealso{
\code{\link{Array}}, \code{\link{Integer}}, \code{\link{EPSGCode}}, \code{\link{String}}, \code{\link{Number}}, 
\code{\link{Date}}, \code{\link{RasterCube}}, \code{\link{VectorCube}}, \code{\link{CallbackArgument}}, 
\code{\link{CallbackValue}}, \code{\link{Variable}}, \code{\link{OutputFormatOptions}}, \code{\link{GeoJson}},
\code{\link{Boolean}}, \code{\link{DateTime}}, \code{\link{Time}}, \code{\link{BoundingBox}}, \code{\link{Kernel}}, 
\code{\link{TemporalInterval}}, \code{\link{TemporalIntervals}}, \code{\link{CollectionId}}, \code{\link{OutputFormat}},
\code{\link{AnyOf}} and \code{\link{ProjDefinition}}
}
