% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{client_version}
\alias{client_version}
\title{Returns the client version}
\usage{
client_version()
}
\arguments{
\item{con}{an OpenEO client}
}
\value{
the client version
}
\description{
The function returns the client version.
}
