% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph_building.R
\name{collection}
\alias{collection}
\title{A collection object}
\usage{
collection(name, ...)
}
\arguments{
\item{name}{the id of the product}

\item{...}{additional parameters that are passed on to get_collection process}
}
\value{
a list represenation for the get_data process
}
\description{
Creates the get_collection process to load data into the process graph.
}
