% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_metadata.R
\name{conformance}
\alias{conformance}
\title{OGC conformance}
\usage{
conformance(con = NULL)
}
\arguments{
\item{con}{a connected openEO client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\description{
Queries the openEO service for the conformance. As stated in the API it is highly optional and  only available if the service
wants to achieve full compatibility with OGC API clients. This function queries the /conformance endpoint and returns it results
as a list object translated from JSON using the jsonlite package.
}
