% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{connect}
\alias{connect}
\title{Connect to a openeEO back-end}
\usage{
connect(host, version = NULL, user = NULL, password = NULL,
  disable_auth = FALSE, auth_type = "bearer", login_type = "basic")
}
\arguments{
\item{host}{URL pointing to the openEO server back-end host}

\item{version}{the version number as string}

\item{user}{the user name (optional)}

\item{password}{the password (optional)}

\item{disable_auth}{flag to specify if the back-end supports authorization on its endpoints}

\item{auth_type}{the general authentication method used on all endpoints. Either "bearer" or "basic".}

\item{login_type}{either "basic" or "oidc". This refers to the login mechanism that shall be used}
}
\description{
Connects to openEO back-end. If the backend provides a well-known endpoint that allows for redirecting to
specific versions, then you should provide the versions parameter.
}
