% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph.R
\name{create_process_graph}
\alias{create_process_graph}
\title{Stores a graph on the back-end}
\usage{
create_process_graph(
  con = NULL,
  graph,
  id = NULL,
  summary = NULL,
  description = NULL,
  submit = TRUE
)
}
\arguments{
\item{con}{connected and authorized openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}

\item{graph}{a process graph definition}

\item{id}{the title of the process graph (optional)}

\item{summary}{the summary of the process graph (optional)}

\item{description}{the description of a process graph (optional)}

\item{submit}{whether to create a process_graph at the service or to create it for local use}
}
\value{
a list assembling a process graph description or the graph id if send
}
\description{
Uploads the process graph information to the back-end and stores it for reuse.
}
