% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{create_service}
\alias{create_service}
\title{Prepares and publishes a service on the back-end}
\usage{
create_service(
  type,
  graph,
  title = NULL,
  description = NULL,
  enabled = NULL,
  configuration = NULL,
  plan = NULL,
  budget = NULL,
  con = NULL
)
}
\arguments{
\item{type}{character the ogc web service type name to be created or an object of type ServiceType, which can be obtained by list_service_types()}

\item{graph}{A \code{\link{Graph}}, a function returning a \code{\link{ProcessNode}} as an endpoint or the \code{\link{ProcessNode}} 
will return the results}

\item{title}{character (optional) the title in human readabled form for the service}

\item{description}{character (optional) the description for the service}

\item{enabled}{logical}

\item{configuration}{a named list specifying the configuration parameter}

\item{plan}{character the billing plan}

\item{budget}{numeric the amount of credits that can be spent for this service}

\item{con}{connected and authenticated openeo clien object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
Service object
}
\description{
The function will send a configuration object to the back-end to create a webservice from a job considering
additional parameter.
}
