% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf.R
\name{create_udf}
\alias{create_udf}
\title{Defines a UDF on the server}
\usage{
create_udf(process, con, prior.name = "collections", language, type, content,
  target, ...)
}
\arguments{
\item{con}{The authorized Connection}

\item{prior.name}{The parameter name of the predecessor of this pipe}

\item{language}{The programming language of the uploaded script}

\item{type}{The udf type}

\item{content}{The local file path of a script the user wants to upload}

\item{target}{The relative path on the users workspace on the openEO back-end}

\item{...}{additional udf parameters}
}
\value{
A named list that represents an UDF as list for the process graph
}
\description{
This function will allow the user to define and uploads the script (content) into
the users workspace (target)
}
