% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph_building.R
\name{create_variable}
\alias{create_variable}
\title{Creates a variable in a process graph}
\usage{
create_variable(
  name,
  description = NULL,
  type = "string",
  subtype = NULL,
  default = NULL
)
}
\arguments{
\item{name}{the name of the variable}

\item{description}{an optional description of the variable}

\item{type}{the type of the value that is replaced on runtime, default 'string'}

\item{subtype}{the subtype of the type (as specified by openEO types)}

\item{default}{the default value for this variable}
}
\value{
a \code{\link{ProcessGraphParameter}} object
}
\description{
This function will create a variable to be used in the designated process graph with additional optional information.
}
