% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{credits}
\alias{credits}
\title{Returns the users available credits}
\usage{
credits(con)
}
\arguments{
\item{con}{connected and authenticated openeo client object}
}
\value{
the credits
}
\description{
Queries the server and returns the available credits of the current user.
}
