% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\name{debug}
\alias{debug}
\alias{is.debugging}
\title{Triggers debugging mode}
\usage{
debug(status = TRUE)

is.debugging()
}
\arguments{
\item{status}{logical Turn on debugging on = TRUE, off = FALSE, default TRUE}
}
\description{
The debugging mode is created to investigate the communication between server and client. The modus can be
turned on or off, depending on the selected status (TRUE, FALSE). It is stored as a dynamical variable in the
package and other package functions can access it naturally.
}
