% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{defineService}
\alias{defineService}
\title{Prepares and publishes a service on the back-end}
\usage{
defineService(con, job_id, service_type, ...)
}
\arguments{
\item{con}{connected and authenticated openeo clien object}

\item{job_id}{the job id to create the service from}

\item{service_type}{one of the supported services (see services())}

\item{...}{additional parameter which are send as 'service_args'}
}
\value{
service representation as list
}
\description{
The function will send a configuration object to the back-end to create a webservice from a job considering
additional parameter.
}
