% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{delete_service}
\alias{delete_service}
\title{Deletes a service function for a job}
\usage{
delete_service(service, con = NULL)
}
\arguments{
\item{service}{the Service or its id}

\item{con}{connected and authorized openEO client object (optional) otherwise \code{\link[=active_connection]{active_connection()}}
is used.}
}
\description{
Queries the back-end and removes the current set service function of job.
}
