% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_defined_processes.R
\name{delete_user_process}
\alias{delete_user_process}
\title{Deletes a user process}
\usage{
delete_user_process(id, con = NULL)
}
\arguments{
\item{id}{the id of the user process}

\item{con}{connected and authorized openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\description{
The function initiates the deletion of a user defined process on the back-end. Only the owning user can delete
their process. The user defined process also should not be part of any particular job.
}
