% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{deleteService}
\alias{deleteService}
\title{Deletes a service function for a job}
\usage{
deleteService(con, service_id)
}
\arguments{
\item{con}{connected and authorized openeo client object}

\item{service_id}{the service id}
}
\description{
Queries the backend and removes the current set service function of job.
}
