% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{describe}
\alias{describe}
\title{Describe a process or product}
\usage{
describe(con, process_id = NA, product_id = NA, ...)
}
\arguments{
\item{con}{Authentication object}

\item{process_id}{id of a process to be described}

\item{product_id}{id of a product to be described}
}
\value{
a list of detailed information
}
\description{
retrieve more detailed information data and processes
}
