% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{describe_account}
\alias{describe_account}
\title{Retrieves the current users account information}
\usage{
describe_account(con = NULL)
}
\arguments{
\item{con}{authenticated client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
object of type user
}
\description{
Calls endpoint \code{/me} to fetch the user account information of the user that is currently logged in to the back-end
}
