% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{describe_job}
\alias{describe_job}
\title{Fetches information about a job}
\usage{
describe_job(job, con = NULL)
}
\arguments{
\item{job}{the job object or the id of the job}

\item{con}{authenticated Connection (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
a detailed description about the job
}
\description{
Returns a detailed description about a specified job (e.g., the status)
}
