% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{describe_process}
\alias{describe_process}
\title{Describe a process}
\usage{
describe_process(con, id = NA)
}
\arguments{
\item{con}{Authentication object}

\item{id}{id of a process to be described}
}
\value{
a list of detailed information
}
\description{
Queries an openeo back-end and retrieves more detailed information about offered processes
}
