% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{describe_service}
\alias{describe_service}
\title{Describes a service}
\usage{
describe_service(service, con = NULL)
}
\arguments{
\item{service}{the Service object or its id}

\item{con}{connected and authorized openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
Service object
}
\description{
Queries the server and returns information about a particular service
}
