% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_defined_processes.R
\name{describe_user_process}
\alias{describe_user_process}
\title{Fetches the representation of a stored user defined process}
\usage{
describe_user_process(id, con = NULL)
}
\arguments{
\item{id}{The id of the user process at the back-end}

\item{con}{connected and authenticated openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
the user process as a ProcessInfo class (list object)
}
\description{
The function queries the back-end for a specific user defined process and returns detailed information.
}
