% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{describeGraph}
\alias{describeGraph}
\title{Fetches the representation of a stored graph}
\usage{
describeGraph(con, graph_id, user_id = NULL)
}
\arguments{
\item{con}{connected and authenticated openeo client object}

\item{graph_id}{The id of a process graph on the back-end}

\item{user_id}{(optional) the user id from which user to fetch the process graph}
}
\value{
the process graph as list
}
\description{
The function queries the back-end for a specific user defined process graph
}
