% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{describeProcess}
\alias{describeProcess}
\title{Describe a process}
\usage{
describeProcess(con, process_id = NA)
}
\arguments{
\item{con}{Authentication object}

\item{process_id}{id of a process to be described}
}
\value{
a list of detailed information
}
\description{
Queries an openeo back-end and retrieves more detailed information about offered processes
}
