% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection-functions.R
\name{dimensions.Collection}
\alias{dimensions.Collection}
\title{Returns dimension information}
\usage{
\method{dimensions}{Collection}(x, ...)
}
\arguments{
\item{x}{a Collection object}

\item{...}{parameters to pass on (not used)}
}
\value{
dimension information as list
}
\description{
The function returns the dimension information of a Collection object. This object is usually obtained when calling
\link{describe_collection}. It returns the meta data information for the cube dimensions.
}
