% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection-functions.R
\name{dimensions.CollectionInfo}
\alias{dimensions.CollectionInfo}
\title{Returns dimension information}
\usage{
\method{dimensions}{CollectionInfo}(x, ...)
}
\arguments{
\item{x}{a CollectionInfo object}

\item{...}{parameters to pass on (not used)}
}
\value{
dimension information as list
}
\description{
The function returns the dimension information of a CollectionInfo object. This object is usually obtained when calling 
\link{describe_collection}. It returns than the meta data information for the cube dimensions.
}
