% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection-functions.R
\name{dimensions}
\alias{dimensions}
\title{Returns dimension}
\usage{
dimensions(x, ...)
}
\arguments{
\item{x}{an object from which dimension information is returned}

\item{...}{additional parameters to pass on to internal functions}
}
\value{
dimension information as list
}
\description{
Returns dimension
}
