% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{disconnect}
\alias{disconnect}
\title{disconnect}
\usage{
disconnect()
}
\value{
invisible NULL
}
\description{
Uses the connections disconnect method to logout and clear all variables in the package for this active back-end. This will also refresh
RStudios connection observer if it can be found.
}
