% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{download_file}
\alias{download_file}
\title{Download a file from the user workspace}
\usage{
download_file(src, dst = NULL, con = NULL)
}
\arguments{
\item{src}{the relative file path of the source file on the openEO back-end}

\item{dst}{the destination file path on the local file system}

\item{con}{authorized connection (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
The file path of the stored file on your machine
}
\description{
Sends a request to an openEO back-end to access the users files and downloads them to a given location.
}
