% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{download_results}
\alias{download_results}
\title{Downloads the results of a job}
\usage{
download_results(job, folder, con = NULL)
}
\arguments{
\item{job}{job object or the job_id for which the results are fetched. Also the return value of
\code{\link[=list_results]{list_results()}} or its 'assets' field is also accepted.}

\item{folder}{a character string that is the target path on the local computer}

\item{con}{a connected and authenticated openEO connection (optional) otherwise \code{\link[=active_connection]{active_connection()}}
is used.}
}
\value{
a list of the target file paths or NULL if 'job' was incorrect
}
\description{
The function will fetch the results of a asynchronous job and will download all files stated in the links. The parameter
'folder' is the target location on the local computer.
}
