% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{downloadJob}
\alias{downloadJob}
\title{Downloads the result of job}
\usage{
downloadJob(con, job_id, format)
}
\arguments{
\item{con}{authenticated Conenction}

\item{job_id}{id of the job}

\item{format}{specification about the format for the result}
}
\description{
Download the result of a finished job
}
