% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{executeTask}
\alias{executeTask}
\title{Executes a job directly and returns the data immediately}
\usage{
executeTask(con, task, format, output)
}
\arguments{
\item{con}{Connection}

\item{task}{A Process or chained processes to a Task}

\item{format}{The inteded format of the data to be returned}

\item{output}{the path, filename and extension, where to store the data}
}
\value{
a connection to file if output was provided, the raw data if not
}
\description{
Executes a job directly on the connected openEO backend and returns the data. It relates to
POST /api/jobs?evaluation="sync" in v0.0.1
POST /api/execute?format=GTiff in v0.0.2
}
