% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{follow_job}
\alias{follow_job}
\title{Follow an executed Job}
\usage{
follow_job(con = NULL, job_id)
}
\arguments{
\item{con}{An authenticated connection (optional) otherwise \code{\link{active_connection}}
is used.}

\item{job_id}{the id of the job on the server the user wants to connect to}
}
\value{
a WebSocket connection
}
\description{
Opens up a websocket to the openEO back-end to fetch updates about a running job.
}
