% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{followJob}
\alias{followJob}
\title{Follow an executed Job}
\usage{
followJob(con, job_id)
}
\arguments{
\item{con}{An authenticated connection}

\item{job_id}{the id of the job on the server the user wants to connect to}
}
\value{
a WebSocket connection
}
\description{
Opens up a websocket to the openEO backend to fetch updates about a running job.
}
