% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{get_job_by_id}
\alias{get_job_by_id}
\title{Fetches information about a job}
\usage{
get_job_by_id(con, id)
}
\arguments{
\item{con}{authenticated Connection}

\item{id}{id of the job}
}
\value{
a detailed description about the job
}
\description{
Returns a detailed description about a specified job. For example to check the status of a job.
}
