% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{get_process_graph_by_id}
\alias{get_process_graph_by_id}
\title{Fetches the representation of a stored graph}
\usage{
get_process_graph_by_id(con, id)
}
\arguments{
\item{con}{connected and authenticated openeo client object}

\item{id}{The id of a process graph on the back-end}
}
\value{
the process graph as list
}
\description{
The function queries the back-end for a specific user defined process graph
}
