% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{get_service_by_id}
\alias{get_service_by_id}
\title{Describes a service}
\usage{
get_service_by_id(con, id)
}
\arguments{
\item{con}{connected and authorized openeo client object}

\item{service_id}{the service id}
}
\value{
service as a list
}
\description{
Queries the server and returns information about a particular service
}
