% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_defined_processes.R
\name{get_user_process}
\alias{get_user_process}
\title{Fetches the representation of a stored user defined process}
\usage{
get_user_process(id, con = NULL)
}
\arguments{
\item{id}{The id of a process graph on the back-end}

\item{con}{connected and authenticated openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
the process graph as list
}
\description{
The function queries the back-end for a specific user defined process and returns detailled information.
}
