% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{graphToJSON}
\alias{graphToJSON}
\title{Wrapper for toJSON}
\usage{
graphToJSON(graph)
}
\arguments{
\item{graph}{a list / nested list representing the process graph}
}
\value{
JSON string of the process graph as a character string
}
\description{
This function is intended to have a preconfigured toJSON function
to allow a user to visualize the process graph in JSON (like it will
be sent to the back-end)
}
