% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{IAuth}
\alias{IAuth}
\title{IAuth}
\description{
An interface that states the intended behavior for the authentication.
}
\section{Fields}{

\describe{
\item{\code{access_token}}{The access_token to query password restricted web services of an openeo back-end}
}}

\section{Methods}{

\describe{
  \item{\code{$login()}}{Initiates the authentication / login in order to obtain the access_token}
  \item{\code{$logout()}}{Terminates the access_token session and logs out the user on the openeo back-end}
}
}

\seealso{
\code{\link{BasicAuth}} or \code{\link{OIDCAuth}}
}
