% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{jobResult}
\alias{jobResult}
\title{Creates a list of download paths}
\usage{
jobResult(con, job_id, format = NULL)
}
\arguments{
\item{con}{connected and authenticated openeo client object}

\item{job_id}{the id of the job}

\item{format}{optional format of the output result, in case it shall differ from the one stated in the job}
}
\value{
vector of URLs
}
\description{
The function queries the backend to receive the URLs to the downloadable files of a particular job.
}
