% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_metadata.R
\name{list_file_formats}
\alias{list_file_formats}
\title{Supported Input/Output formats}
\usage{
list_file_formats(con = NULL)
}
\arguments{
\item{con}{openEO client object (optional) otherwise \code{\link[=active_connection]{active_connection()}}
is used.}
}
\value{
a FileFormatList object
}
\description{
The function queries the openEO service for supported I/O formats as a \code{FileFormatList} object.
}
\details{
The \code{FileFormatList} object is a named list, which is organized into 'input' and 'output'. For each category a different
named list with the \code{FileFormat} is indexed by its format ID.
}
