% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_metadata.R
\name{list_file_types}
\alias{list_file_types}
\title{Returns the output formats}
\usage{
list_file_types(con)
}
\arguments{
\item{con}{openeo client object}
}
\value{
a data frame with formats, the applied output data type ('raster', 'vector', 'table' and/or 'other') and optional configuration parameter
}
\description{
The function queries the back-end for supported output formats.
}
