% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{list_files}
\alias{list_files}
\title{Lists workspace files}
\usage{
list_files(con)
}
\arguments{
\item{con}{authorized connection}
}
\value{
a tibble of for filenames and their sizes
}
\description{
Lists all files in the workspaces of the authenticated user.
}
