% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{list_jobs}
\alias{list_jobs}
\title{List the jobs of a user}
\usage{
list_jobs(con = NULL)
}
\arguments{
\item{con}{the authenticated Connection (optional) otherwise \code{\link[=active_connection]{active_connection()}}
is used.}
}
\description{
Lists the jobs that a user has uploaded or in that are in execution
}
