% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{list_oidc_providers}
\alias{list_oidc_providers}
\title{Available OIDC provider}
\usage{
list_oidc_providers(con = NULL)
}
\arguments{
\item{con}{active openEO service connection (\code{\link{OpenEOClient}})}
}
\value{
a \code{ProviderList} object which is a named list of \code{Provider} objects.
}
\description{
In case the openEO service provider supports OpenID connect authentication, this function will return a list
of supported provider that can be used on this specific service.
}
