% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{list_results}
\alias{list_results}
\title{Creates a list of download paths}
\usage{
list_results(con, job)
}
\arguments{
\item{con}{connected and authenticated openeo client object}

\item{job}{the job object or the id of the job}
}
\value{
result object containing of URLs for download
}
\description{
The function queries the back-end to receive the URLs to the downloadable files of a particular job.
}
