% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{list_service_types}
\alias{list_service_types}
\title{Returns the offered webservice types of the back-end}
\usage{
list_service_types(con)
}
\arguments{
\item{con}{a connected openeo client object}
}
\value{
vector of identifier of supported webservice
}
\description{
The function queries the back-end for the supported webservice types that can be used on the client.
}
