% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{list_services}
\alias{list_services}
\title{Lists the current users services}
\usage{
list_services(con = NULL)
}
\arguments{
\item{con}{connected and authenticated openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
list of services lists
}
\description{
Queries the back-end to retrieve a list of services that the current user owns. Services are 
webservices like WCS, WFS, etc.
}
