% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{list_udf_runtimes}
\alias{list_udf_runtimes}
\title{Lists the supported UDF runtimes}
\usage{
list_udf_runtimes(con)
}
\arguments{
\item{con}{connected and authenticated openeo client object}
}
\value{
list of udf runtimes with supported udf types, versions and installed packages
}
\description{
The function queries the back-end for its supported udf runtimes and returns detailed information about each
runtime.
}
