% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_defined_processes.R
\name{list_user_processes}
\alias{list_user_processes}
\title{Lists the Ids of the process graphs from the current user}
\usage{
list_user_processes(con = NULL)
}
\arguments{
\item{con}{connected and authenticated openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
a named list of user defined processes (\code{ProcessInfo})
}
\description{
Queries the back-end to retrieve a list of graph ids that the current user has stored on the back-end.
}
