% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{listCollections}
\alias{listCollections}
\title{List Data on conected server}
\usage{
listCollections(con)
}
\arguments{
\item{con}{Connection object}
}
\description{
List available collections stored on a openEO server
}
