% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{listEndpoints}
\alias{listEndpoints}
\title{Shows an overview about the capabilities of an OpenEO back-end}
\usage{
listEndpoints(con)
}
\arguments{
\item{con}{A connected OpenEO client}
}
\value{
tibble
}
\description{
The client queries the version resolved back-end for its endpoint capabilities and returns it as
a tibble.
}
