% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{listFormats}
\alias{listFormats}
\title{Returns the output formats}
\usage{
listFormats(con)
}
\arguments{
\item{connected}{openeo client object}
}
\value{
list of formats with optional configuration parameter
}
\description{
The function queries the backend for supported output formats.
}
