% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{listGraphs}
\alias{listGraphs}
\title{Lists the Ids of the process graphs from the current user}
\usage{
listGraphs(con)
}
\arguments{
\item{con}{connected and authenticated openeo client object}
}
\value{
vector of process graph ids
}
\description{
Queries the back-end to retrieve a list of graph ids that the current user has stored on the back-end.
}
