% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{listJobs}
\alias{listJobs}
\title{List the jobs that a user has}
\usage{
listJobs(con)
}
\arguments{
\item{con}{the authenticated Connection}
}
\description{
lists the jobs that a user has uploaded or in execution
}
