% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{listProcesses}
\alias{listProcesses}
\title{List available processes on server}
\usage{
listProcesses(con)
}
\arguments{
\item{con}{Connection object}
}
\value{
a list of lists with process_id and description
}
\description{
list all processes available on the back-end
}
