% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{listResults}
\alias{listResults}
\title{Creates a list of download paths}
\usage{
listResults(con, job_id)
}
\arguments{
\item{con}{connected and authenticated openeo client object}

\item{job_id}{the id of the job}
}
\value{
result object containing of URLs for download
}
\description{
The function queries the back-end to receive the URLs to the downloadable files of a particular job.
}
