% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{listServices}
\alias{listServices}
\title{Lists the current users services}
\usage{
listServices(con)
}
\arguments{
\item{con}{connected and authenticated openeo client object}
}
\value{
list of services lists
}
\description{
Queries the backend to retrieve a list of services that the current user owns. Services are 
webservices like WCS, WFS, etc.
}
