% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{login}
\alias{login}
\title{Function to login to a specific backend}
\usage{
login(con, user = NULL, password = NULL, login_type = NULL,
  external = NULL)
}
\arguments{
\item{con}{connected back-end connection}

\item{user}{the user name}

\item{password}{the password}

\item{login_type}{either NULL, 'basic' or 'oidc'. This refers to the login mechanism that shall be used. NULL disables authentication.}

\item{external}{character - 'google' whether Google is used as a Identity Provider for OIDC}
}
\value{
a connected and authenticated back-end connection
}
\description{
Retrieves the bearer-token from the backend by sending user name and password to the backend. This step
is usually also performed in the 'connect' step. But if you only connected to a back-end in order to 
register, then you need to log in afterwards.
}
\examples{
\dontrun{
# simple connection without login to maybe explore the capabilities of a back-end first
con = connect(host='http://example.openeo.org',version='0.4.2')

login(con=con, user='user',password='password',login_type='basic')

# or alternatively the oidc login
login(con=con,login_type='oidc')
}
}
