% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{logout}
\alias{logout}
\title{Logout}
\usage{
logout(con = NULL)
}
\arguments{
\item{con}{a connected openEO client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\description{
Logs out or closes the active connection to an openEO service.
}
