% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\name{logs}
\alias{logs}
\title{Access logs of a Service or Job}
\usage{
logs(obj = NULL, job_id = NULL, service_id = NULL, con = NULL, timeout = 10)
}
\arguments{
\item{obj}{Service or Job object}

\item{job_id}{character the jobs ID}

\item{service_id}{character - the services ID}

\item{con}{a connected openeo client (optional) otherwise \code{\link{active_connection}}
is used.}

\item{timeout}{integer the timeout for the logging after no update in seconds, default is 10}
}
\description{
Prints contents of the log file of a Job or Service to the console. Requests the log every second. If the log response always empty 
for a given timeout, the logging stops.
}
