% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-functions.R
\name{map_collection}
\alias{map_collection}
\title{Creates a map of collection footprints}
\usage{
map_collection(con, collection)
}
\arguments{
\item{con}{A connected openeo back-end connection}

\item{collection}{character one string or a vector of strings with valid collection ids}
}
\value{
a mapview object containing footprints and baselayers
}
\description{
This function uses a back-end connection and one or more collection ids and prints their footprints
on a mapview map.
}
