% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-functions.R
\name{map_service}
\alias{map_service}
\title{Creates a map with service}
\usage{
map_service(con, service_id, layers, transparent = TRUE)
}
\arguments{
\item{con}{a authenticated connection to an openeo back-end}

\item{service_id}{The service id which is going to be used}

\item{layers}{A character string containing the layers that shall be shown}

\item{transparent}{logical if the "free" area shall be transparent}
}
\value{
a leaflet map
}
\description{
The function uses leaflet as a map component and having an OSM layer as background and stacks the
OGC services on top.
}
