% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_viewer.R
\name{markdownViewer}
\alias{markdownViewer}
\title{Function to use RStudio viewer to view markdown text}
\usage{
markdownViewer(mdtext, isgithub = FALSE)
}
\arguments{
\item{mdtext}{markdown text}

\item{isgithub}{logical parameter to state whether Github markdown was used or not. Default: FALSE}
}
\description{
It is intended to view data or process description in a human readable way. Therefore markdown will be allowed in the respective `description`
fields. This function provides means to visualize the Markdown code as HTML in a RStudio viewer.
}
\examples{
\dontrun{
con = connect(host='http://example.openeo.org/v/0.4.2',
              user='user',
              password='password',
              login_type='basic')

collection = describe_collection(con = con, id = 'some_collection_id')

# the description of a collection is allowed to use markdown
markdownViewer(collection$description)
}

}
