% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{modifyService}
\alias{modifyService}
\title{Modifies a service}
\usage{
modifyService(con, service_id, ...)
}
\arguments{
\item{con}{connected and authorized openeo client object}

\item{service_id}{the service id}
}
\description{
Modifies a service
}
