% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R, R/utilities.R
\name{openeo-deprecated}
\alias{openeo-deprecated}
\alias{graphToJSON}
\alias{processToJSON}
\title{openeo-deprecated}
\usage{
graphToJSON(x, ...)

processToJSON(x, ...)
}
\description{
Lists all currently deprecated functions that will be removed in the future.
}
