% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{data}
\name{OpenEOClient}
\alias{OpenEOClient}
\title{OpenEO client class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OpenEOClient
}
\description{
A R6Class that interacts with an openEO-conformant backend.
}
\keyword{datasets}
