% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{orderResult}
\alias{orderResult}
\title{Executes a job directly and stores it on the server}
\usage{
orderResult(con, task, format, path)
}
\arguments{
\item{con}{Connection}

\item{task}{A Process or chained processes to a Task}

\item{format}{The inteded format of the data to be returned}

\item{path}{the relative path in the users workspace, where to store the data}
}
\value{
A named list or vector with "job_id" and "path" to the file in the users workspace
}
\description{
Executes a job on a server directly and stores the data in the user workspace
on the server. It relates to POST /api/jobs?evaluation="batch"
}
