% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{pauseJob}
\alias{pauseJob}
\title{Pauses a running batch job}
\usage{
pauseJob(con, job_id)
}
\arguments{
\item{con}{authenticated Connection}

\item{job_id}{id of job that will be canceled}
}
\value{
a success / failure notification
}
\description{
Sends a request to the server in order to pause a running batch job.
}
