% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{pgb}
\alias{pgb}
\title{Get a process graph builder from the connection}
\usage{
pgb(con)
}
\arguments{
\item{con}{a connection to an openeo back-end}
}
\value{
a ProcessGraphBuilder class with the offered processes of the backend
}
\description{
Queries the connected back-end for all available processes and registers them as R functions on
a process builder class to build up the process graph.
}
