% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_metadata.R
\name{privacy_policy}
\alias{privacy_policy}
\title{Visualizes the privacy policy}
\usage{
privacy_policy(con = NULL)
}
\arguments{
\item{con}{a connected openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
a list of the link identifying the privacy policy from the service capabilities or NULL
}
\description{
If the service provides information in their capabilities about their privacy policy, the function opens a new RStudio 
viewer panel and visualizes the HTML content of the link.
}
