% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{process}
\alias{process}
\title{Create a process object}
\usage{
process(process, process_id, ...)
}
\arguments{
\item{process}{A process that might be chained with this process}

\item{process_id}{ID of the process offered by the connected openEO backend}

\item{...}{named arguments that are passed to the process description}
}
\description{
defines a process with arguments
}
