% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{process_viewer}
\alias{process_viewer}
\title{Viewer panel for provided openEO processes}
\usage{
process_viewer(x, con = NULL)
}
\arguments{
\item{x}{a function from the \code{\link{ProcessCollection}}, a \code{\link{ProcessNode}},
\code{\link{Process}} or a character containing the process id.}

\item{con}{a specific connection (optional), last connected service if omitted.}
}
\description{
Opens up a viewer panel in RStudio and renders one or more processes of the connected 
openEO service in HTML. The components of openeo-js-commons / openEO webeditor are reused.
}
