% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessGraphBuilder-class.R
\docType{data}
\name{ProcessGraphBuilder}
\alias{ProcessGraphBuilder}
\title{Process graph builder}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProcessGraphBuilder
}
\description{
This class shall support the proper definition of process graphs by registering all supported processes of the backend
to which you are connected on this class. On initialization the class will be extended with processes that are offered. As
basic function the function 'process' provided by this package will be wrapped and the ... parameter will be replaced by
NULL initialized parameter of the arguments in the backends detailed process description.
}
\examples{
\dontrun{
con = connect(host = "http://localhost:8000/api/",user = "test",password = "test",rbackend=TRUE)
pg = ProcessGraphBuilder$new(con)
collection(collection_id = "Sentinel_2A",id_name = "product_id") \%>\% pg$filter_bands(bands = "B04")
}
}
\keyword{datasets}
