% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{ProcessGraphParameter}
\alias{ProcessGraphParameter}
\title{ProcessGraphParameter}
\value{
Object of \code{\link{R6Class}} which represents a ProcessGraph value.
}
\description{
Inheriting from \code{\link{Argument}} in order to represent the available data within a ProcessGraph graph.
Hints from the openeo api documention:
\itemize{
  \item \url{https://open-eo.github.io/openeo-api/processes/#callbacks}
  \item \url{https://open-eo.github.io/openeo-api/v/0.4.2/processgraphs/#callbacks}
}
}
\seealso{
\code{\link{Array}}, \code{\link{Integer}}, \code{\link{EPSGCode}}, \code{\link{String}}, \code{\link{Number}}, 
\code{\link{Date}}, \code{\link{RasterCube}}, \code{\link{VectorCube}}, \code{\link{ProcessGraph}}, 
\code{\link{ProcessGraphParameter}}, \code{\link{OutputFormatOptions}}, \code{\link{GeoJson}},
\code{\link{Boolean}}, \code{\link{DateTime}}, \code{\link{Time}}, \code{\link{BoundingBox}}, \code{\link{Kernel}}, 
\code{\link{TemporalInterval}}, \code{\link{TemporalIntervals}}, \code{\link{CollectionId}}, \code{\link{OutputFormat}},
\code{\link{AnyOf}} and \code{\link{ProjDefinition}}
}
