% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph_building.R
\name{ProcessNode}
\alias{ProcessNode}
\title{Process Node object}
\description{
This class inherits all functions and fields from \code{\link{Process}} and extends it with a node id and a 
special serialization function. The ProcessNode is an essential building block of the \code{\link{Graph}}.
}
\section{Methods}{

\describe{
   \item{$getNodeId()}{returns the node id}
   \item{$setNodeId(id)}{set the node id, which is of interest when \code{\link{parse_graph}} is executed}
   \item{$serializeAsReference()}{during the serialization the process node might be used as a 
   reference and this function serializes the process node accordingly}
}
}

\section{Arguments}{

\describe{
   \item{id}{the node id}
}
}

