% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{processToJSON}
\alias{processToJSON}
\title{Wrapper for toJSON}
\usage{
processToJSON(x, ...)
}
\arguments{
\item{process}{a list / nested list representing the process}
}
\value{
JSON string of the process as a character string
}
\description{
This function is intended to have a preconfigured toJSON function
to allow a user to visualize the process in JSON (like it will
be sent to the back-end)
}
