% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{queryJob}
\alias{queryJob}
\title{Fetches information about a job}
\usage{
queryJob(con, job_id)
}
\arguments{
\item{con}{authenticated Connection}

\item{job_id}{id of the job}
}
\value{
a detailed description about the job
}
\description{
Returns a detailed description about a specified job. For example to check the status of a job.
}
