% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{queueJob}
\alias{queueJob}
\title{Queues a job on the back-end for evaluation}
\usage{
queueJob(con, job_id)
}
\arguments{
\item{con}{connected and authenticated openeo client}

\item{job_id}{the job id of a created job}
}
\description{
The function will queue a job for evaluation that has been created before by 'createJob'. There will be
a notification for success or failure, but no return value.
}
