% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{queueTask}
\alias{queueTask}
\title{Stores a job on the backend for execution on demand}
\usage{
queueTask(con, task)
}
\arguments{
\item{con}{Connection}

\item{task}{A Process or chained processes to a Task}

\item{format}{The inteded format of the data to be returned}
}
\description{
Uploads a job to a server for lazy evaluation
on the server. It relates to POST /api/jobs?evaluation="batch"
}
