% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{RasterCube-deprecated}
\alias{RasterCube-deprecated}
\title{RasterCube}
\value{
Object of \code{\link[R6:R6Class]{R6::R6Class()}} representing a raster cube.
}
\description{
Inheriting from \link{DataCube} in order to represent a raster cube. This is usually the in- and
output format of a process unless the process operates within a ProcessGraph on reduced data.
The \code{\link[=VectorCube-deprecated]{VectorCube-deprecated()}} behaves comparably, but with underlying spatial feature data.
}
\seealso{
\code{\link[=openeo-deprecated]{openeo-deprecated()}}
}
