% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph_building.R
\name{remove_variable}
\alias{remove_variable}
\title{Removes a variable from the Graph}
\usage{
remove_variable(graph, variable)
}
\arguments{
\item{graph}{a \code{\link{Graph}} object}

\item{variable}{a variable id or a variable object}
}
\value{
TRUE
}
\description{
The function that removes a selected variable from the graph. It is removed from the list of defined 
variables that are obtainable with \code{\link{variables}}. The variables already placed in the graph 
won't be deleted, only in the defined variables list.
}
