% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{replaceGraph}
\alias{replaceGraph}
\title{Replaces the current graph with a given}
\usage{
replaceGraph(con, graph_id, graph)
}
\arguments{
\item{con}{connected and authorized openeo client object}

\item{graph_id}{process graph id}

\item{graph}{a process graph definition created by chaining "process()", "collection()" or using a ProcessGraphBuilder}
}
\description{
Upload a process graph to the back-end under a given (existing) process graph.
}
