% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_bbox.ProcessNode}
\alias{st_bbox.ProcessNode}
\title{st_bbox for ProcessNode}
\usage{
\method{st_bbox}{ProcessNode}(obj, ...)
}
\arguments{
\item{obj}{the process node}

\item{...}{not used}
}
\value{
sf bbox object if one element was found, else a list of all bounding boxes (usually returned in EPSG:4326)
}
\description{
Traverses the graph from end node to roots and searches for defined bounding boxes in load_collection, filter_spatial, filter_bbox.
}
