% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{start_job}
\alias{start_job}
\title{Starts remote asynchronous evaluation of a job}
\usage{
start_job(job, log = FALSE, con = NULL)
}
\arguments{
\item{job}{the job object or the job id of the defined job}

\item{log}{logical - whether to enable automatic logging after starting the job}

\item{con}{connected and authenticated openeo client (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
the job_id of the defined job
}
\description{
The function sends a start signal to the backend in order to start processing the results
for a defined job.
}
