% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/client.R, R/jobs.R, R/services.R
\name{status}
\alias{status}
\alias{status.OpenEOClient}
\alias{status.Job}
\alias{status.Service}
\title{Retrieves the status}
\usage{
status(x, ...)

\method{status}{OpenEOClient}(x, ...)

\method{status}{Job}(x, ...)

\method{status}{Service}(x, ...)
}
\arguments{
\item{x}{an object like Job}

\item{...}{currently not used}
}
\value{
status as character
}
\description{
The function refreshes the passed object and returns its status.
}
