% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{stop_job}
\alias{stop_job}
\title{Terminates a running job}
\usage{
stop_job(job, con = NULL)
}
\arguments{
\item{job}{the job object or the id of job that will be canceled}

\item{con}{authenticated Connection (optional) otherwise \code{\link[=active_connection]{active_connection()}}
is used.}
}
\value{
a success / failure notification
}
\description{
Informs the server that the specified job needs to be terminated to prevent further costs.
}
