% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{storeGraph}
\alias{storeGraph}
\title{Stores a graph on the backend}
\usage{
storeGraph(con, graph)
}
\arguments{
\item{con}{connected and authorized openeo client object}

\item{graph}{a process graph definitions}
}
\description{
Uploads the process graph information to the backend and stores it for reuse.
}
