% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{udfRuntimes}
\alias{udfRuntimes}
\title{Lists the supported UDF runtimes}
\usage{
udfRuntimes(con)
}
\arguments{
\item{con}{connected and authenticated openeo client object}
}
\value{
list of udf runtimes with supported udf types, versions and installed packages
}
\description{
The function queries the backend for its supported udf runtimes and returns detailed information about each
runtime.
}
