% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{UdfRuntimeVersionArgument}
\alias{UdfRuntimeVersionArgument}
\title{UdfRuntimeVersionArgument class}
\value{
Object of \code{\link[R6:R6Class]{R6::R6Class()}} is an argument that expects a UDF runtime version or character as value.
}
\description{
Inheriting from \link{Argument} in order to represent the id of a UDF runtime object as obtainable by \code{\link[=list_udf_runtimes]{list_udf_runtimes()}}.
}
\seealso{
\link{Array}, \link{Integer}, \link{EPSGCode}, \link{String}, \link{Number},
\link{Date}, \link{DataCube}, \link{ProcessGraphArgument},
\link{ProcessGraphParameter}, \link{OutputFormatOptions}, \link{GeoJson},
\link{Boolean}, \link{DateTime}, \link{Time}, \link{BoundingBox}, \link{Kernel},
\link{TemporalInterval}, \link{TemporalIntervals}, \link{CollectionId}, \link{OutputFormat},
\link{AnyOf}, \link{ProjDefinition}, \link{UdfCodeArgument}, \link{UdfRuntimeArgument} and
\link{UdfRuntimeVersionArgument}, \link{MetadataFilter}
}
