% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{update_job}
\alias{update_job}
\title{Modifies a job with given parameter}
\usage{
update_job(con, job_id, title = NULL, description = NULL,
  process_graph = NULL, plan = NULL, budget = NULL, format = NULL, ...)
}
\arguments{
\item{con}{connected and authenticated openeo client}

\item{job_id}{the job id of a created job}

\item{...}{The parameter you want to change. See Details for more information}
}
\description{
The function will mofidy a stored job with the given parameter. The dot parameter will contain all the values
that shall be replaced or removed. Shows a message of result or failure.
}
\details{
The '...' operator shall contain all the values that are to be replaced in the job. There are some reserved
keys. 
'task' will replace the process graph with a newly defined one, therefore the task needs to be a list like in
createJob.
'graph_id' will replace the process graph with the stated process graph id.
'format' will change the desired output format.
All other parameter will be assumed to be special output parameter. Remember, you don't need to specify a task or graph_id,
e.g. if you just want to update the output format. 
To leave parameter unchanged, then don't mention it in the ... parameter. If you want to delete some, then set them to NULL.
}
