% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_defined_processes.R
\name{update_user_process}
\alias{update_user_process}
\title{Update an user defined process}
\usage{
update_user_process(
  id,
  graph = NULL,
  summary = NULL,
  description = NULL,
  con = NULL,
  ...
)
}
\arguments{
\item{id}{process graph id}

\item{graph}{a process graph definition created by combining 'process()', 'collection()' or using a ProcessGraphBuilder}

\item{summary}{summary of the process graph (optional)}

\item{description}{description of the process graph (optional)}

\item{con}{connected and authorized openEO client object (optional) otherwise \code{\link{active_connection}}
is used.}

\item{...}{additional parameters passed to toJSON (like 'digits')}
}
\description{
You can change details on an already created user defined process. You can either edit the meta data like the summary or the description. Or
you can replace the process graph. However, you cannot delete the process graph, but by passing NA to the meta data fields you can empty those
fields in the user defined process.
}
