% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{user.account}
\alias{user.account}
\title{Retrieves the current users account information}
\usage{
user.account(con)
}
\arguments{
\item{con}{authenticated client object}
}
\value{
object of type user
}
\description{
Calls endpoint /me to fetch the user account information of the user that is currently logged in to the back-end
}
