% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_defined_processes.R
\name{user_processes}
\alias{user_processes}
\title{Process collection for user defined processes}
\usage{
user_processes(con = NULL)
}
\arguments{
\item{con}{a connection to an openEO back-end (optional). Otherwise \code{\link{active_connection}}
is used in order to access personal user defined processes. You need to be logged in}
}
\value{
\code{\link{UserProcessCollection}}
}
\description{
The created process graphs via \code{\link{create_user_process}} at the openEO service are user defined processes. 
They can be used for the creation of process graphs themselves. For processes provided by the 
particular openEO service the \code{\link{processes}} function can be used to obtain a builder for those processes. 
Analogous to this idea, this function creates a builder object for user defined processes listed and described
in \code{\link{describe_user_process}} and \code{\link{list_user_processes}}.
}
