% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_defined_processes.R
\name{UserProcessCollection}
\alias{UserProcessCollection}
\title{User Defined Process Collection}
\description{
This object contains template functions from the users stored user defined processes (UDP), which can be reused in other process graphs.
}
\details{
This object is an unlocked R6 object, that allows us to add new functions to this object at runtime. It is structured in the same way 
as the \code{\link{ProcessCollection}} for predefined processes by the openEO back-end. A \code{\link{UserProcessCollection}} is usually created at 
\code{\link{user_processes}}. If you have submitted new user defined processes to the back-end, make sure to call \code{\link{user_processes}} again
to fetch the latest status.
}
\section{Methods}{

\describe{
   \item{\code{$new(con = NULL)}}{The object creator created an openEO connection.} 
}
}

\section{Arguments}{

\describe{
   \item{con}{optional - an active and authenticated Connection (optional) otherwise \code{\link{active_connection}}
is used.}
}
}

