% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_defined_processes.R
\name{validate_process}
\alias{validate_process}
\title{Validate a user process}
\usage{
validate_process(graph, con = NULL, ...)
}
\arguments{
\item{graph}{the process graph that will be sent to the service to be validated}

\item{con}{connected and authorized openEO client object (optional) otherwise \code{\link[=active_connection]{active_connection()}}
is used.}

\item{...}{additional parameters passed to jsonlite::toJSON() (like 'digits')}
}
\description{
Sends the process graph as a user process to the openEO service and validates it with the predefined and user-defined
processes of the service.
}
