% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{validateProcessGraph}
\alias{validateProcessGraph}
\title{Validates a process graph}
\usage{
validateProcessGraph(con, graph)
}
\arguments{
\item{con}{connected and authorized openeo client object}

\item{graph}{the process graph that will be sent to the back-end and is being validated}
}
\description{
Sends the process graph to the back-end and validates it against the offered processes by the backend.
}
