% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{Variable}
\alias{Variable}
\title{Variable class}
\value{
Object of \code{\link{R6Class}} which represents a variable.
}
\description{
This class reflects a variable that can be used within a process graph. It inherits all fields and functions
from \code{\link{Argument}}. If the value is not set, the variable is serialized as a variable, but if it is, 
then the variable is replaced with the set value. The variable is created by \code{\link{create_variable}}.
}
\seealso{
\code{\link{Array}}, \code{\link{Integer}}, \code{\link{EPSGCode}}, \code{\link{String}}, \code{\link{Number}}, 
\code{\link{Date}}, \code{\link{RasterCube}}, \code{\link{VectorCube}}, \code{\link{ProcessGraph}}, 
\code{\link{ProcessGraphParameter}}, \code{\link{Variable}}, \code{\link{OutputFormatOptions}}, \code{\link{GeoJson}},
\code{\link{Boolean}}, \code{\link{DateTime}}, \code{\link{Time}}, \code{\link{BoundingBox}}, \code{\link{Kernel}}, 
\code{\link{TemporalInterval}}, \code{\link{TemporalIntervals}}, \code{\link{CollectionId}}, \code{\link{OutputFormat}},
\code{\link{AnyOf}} and \code{\link{ProjDefinition}}
}
