% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{VectorCube-deprecated}
\alias{VectorCube-deprecated}
\title{VectorCube}
\value{
Object of \code{\link[R6:R6Class]{R6::R6Class()}} representing a vector cube.
}
\description{
Inheriting from \link{DataCube} in order to represent a vector cube. This is analogous to
the \code{\link[=RasterCube-deprecated]{RasterCube-deprecated()}}.
}
\seealso{
\code{\link[=openeo-deprecated]{openeo-deprecated()}}
}
