% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{api_versions}
\alias{api_versions}
\title{Returns the suppported OpenEO API versions}
\usage{
api_versions(url)
}
\arguments{
\item{url}{the url as String pointing to the base host of the back-end}
}
\value{
a tibble containing all supported API versions of the back-end
}
\description{
The function queries the back-end for its supported versions. The endpoint \link[https://open-eo.github.io/openeo-api/apireference/#tag/Capabilities/paths/~1.well-known~1openeo/get]{/.well-known/openeo} 
is called and the JSON result is coerced into a tibble.
}
