% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-functions.R
\name{as.Graph}
\alias{as.Graph}
\alias{as.Graph.ProcessNode}
\alias{as.Graph.function}
\alias{as.Graph.ProcessInfo}
\title{Coercion into Graph}
\usage{
as.Graph.ProcessNode(from)

as.Graph.function(from)

as.Graph.ProcessInfo(from)
}
\arguments{
\item{from}{the source from which to coerce (\code{ProcessNode}, \code{function} or \code{ProcessInfo})}
}
\value{
\code{Graph}
}
\description{
Currently there are several ways to create a \code{Graph} object. One of them is by coercion. You can coerce
from \code{ProcessNode}, a function that returns a ProcessNode or from a userdefined process (\code{ProcessInfo}), as obtained by
function \code{describe_process_graph}.
}
