% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-functions.R
\name{as.Process}
\alias{as.Process}
\alias{as.Process.ProcessInfo}
\alias{as.Process.Graph}
\alias{as.Process.ProcessNode}
\alias{as.Process.Service}
\alias{as.Process.function}
\alias{as.Process.Job}
\title{Coerce into a Process}
\usage{
as.Process.ProcessInfo(from)

as.Process.Graph(from)

as.Process.ProcessNode(from)

as.Process.Service(from)

as.Process.function(from)

as.Process.Job(from)
}
\arguments{
\item{from}{the source from which to coerce (\code{ProcessInfo}, \code{\link{Graph}} or \code{\link{ProcessNode}})}
}
\value{
\code{\link{Process}}
}
\description{
This function converts objects into a process. If no meta data is provided it will return a valid user defined process, not yet storable in the back-end.
}
