% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{BoundingBox}
\alias{BoundingBox}
\title{BoundingBox}
\value{
Object of \code{\link{R6Class}} which represents a bounding box / extent.
}
\description{
Inheriting from \code{\link{Argument}} in order to represent a bounding box / extent of a certain area of 
interest. Its value is usually a named list with "west","south","east" and "north". A NA value means a open
interval.
}
\seealso{
\code{\link{Array}}, \code{\link{Integer}}, \code{\link{EPSGCode}}, \code{\link{String}}, \code{\link{Number}}, 
\code{\link{Date}}, \code{\link{RasterCube}}, \code{\link{VectorCube}}, \code{\link{ProcessGraph}}, 
\code{\link{ProcessGraphParameter}}, \code{\link{OutputFormatOptions}}, \code{\link{GeoJson}},
\code{\link{Boolean}}, \code{\link{DateTime}}, \code{\link{Time}}, \code{\link{BoundingBox}}, \code{\link{Kernel}}, 
\code{\link{TemporalInterval}}, \code{\link{TemporalIntervals}}, \code{\link{CollectionId}}, \code{\link{OutputFormat}},
\code{\link{AnyOf}} and \code{\link{ProjDefinition}}
}
