% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{capabilities}
\alias{capabilities}
\title{Shows an overview about the capabilities of an OpenEO back-end}
\usage{
capabilities(con)
}
\arguments{
\item{con}{A connected OpenEO client}
}
\value{
capabilities object
}
\description{
Queries the back-end for its general capabilities.
}
