% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_metadata.R
\name{capabilities}
\alias{capabilities}
\title{Shows an overview about the capabilities of an OpenEO back-end}
\usage{
capabilities(con = NULL)
}
\arguments{
\item{con}{A connected OpenEO client (optional), if omitted \code{\link{active_connection}} is used}
}
\value{
capabilities object
}
\description{
Queries the back-end for its general capabilities.
}
