% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{capabilities}
\alias{capabilities}
\title{Returns the offered enpoints of the openEO API}
\usage{
capabilities(con)
}
\arguments{
\item{con}{A connected openeo client}
}
\value{
data.frame containing the supported / implemented endpoints of the backend
}
\description{
The function queries the backend for its capabilities. The offered enpoints that are specified in the openeo
API are then returned.
}
