% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_metadata.R
\name{capabilities}
\alias{capabilities}
\title{Capabilities overview}
\usage{
capabilities(con = NULL)
}
\arguments{
\item{con}{A connected OpenEO client (optional), if omitted \code{\link{active_connection}} is used}
}
\value{
capabilities object
}
\description{
The function queries the openEO service to which was connected for general information about the service.
}
