% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{collection}
\alias{collection}
\title{A collection object}
\usage{
collection(collection_id)
}
\arguments{
\item{collection_id}{the id of the product}
}
\value{
a list represenation for a collection / product
}
\description{
creates a list represenation of a collection object
}
