% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph_building.R
\name{collection}
\alias{collection}
\title{A collection object}
\usage{
collection(product_id, id_name = "product_id")
}
\arguments{
\item{product_id}{the id of the product}

\item{id_name}{the name of the identifier for this collection, default "product_id"}
}
\value{
a list represenation for a collection / product
}
\description{
creates a list representation of a collection object
}
