% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{collection_viewer}
\alias{collection_viewer}
\title{View for openEO collections}
\usage{
collection_viewer(x, con = NULL)
}
\arguments{
\item{x}{character with the name of a collection or the \code{\link{Collection}} obtained
with \code{\link{describe_collection}}.}

\item{con}{a specific connection (optional), last connected service if omitted.}
}
\description{
The function opens up a viewer panel in RStudio which renders the collection information
nicely in an HTML. It reuses common components from the openEO webeditor / openeo-js-commons.
}
