% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{compute_result}
\alias{compute_result}
\title{Executes a job directly and returns the data immediately}
\usage{
compute_result(
  graph,
  output_file = NULL,
  budget = NULL,
  plan = NULL,
  con = NULL
)
}
\arguments{
\item{graph}{A process graph}

\item{output_file}{Where to store the retrieved data under}

\item{budget}{numeric, how much to spend at maximum on testing}

\item{plan}{character, selection of a service plan}

\item{con}{connected and authenticated openeo client (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
a connection to file if output was provided, the raw data if not
}
\description{
Executes a job directly on the connected openEO back-end and returns the data. It relates to
POST /api/execute in v0.0.2. During the execution phase the connection to the server remains open.
}
