% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{compute_result}
\alias{compute_result}
\title{Executes a job directly and returns the data immediately}
\usage{
compute_result(con, graph, format = NULL, output_file = NULL, ...)
}
\arguments{
\item{con}{connected and authenticated openeo client}

\item{graph}{A process graph}

\item{format}{The inteded format of the data to be returned}

\item{output_file}{Where to store the retrieved data under}

\item{...}{additional configuration parameter for output generation}
}
\value{
a connection to file if output was provided, the raw data if not
}
\description{
Executes a job directly on the connected openEO back-end and returns the data. It relates to
POST /api/execute in v0.0.2. During the execution phase the connection to the server remains open.
}
