% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{connect}
\alias{connect}
\title{Connect to a openeEO backend}
\usage{
connect(host, user, password)
}
\arguments{
\item{host}{URL pointing to the openEO server backend host}

\item{user}{the user name}

\item{password}{the password}
}
\description{
connects to openEO backend
}
