% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{connect}
\alias{connect}
\title{Connect to a openeEO service}
\usage{
connect(
  host,
  version = NULL,
  user = NULL,
  password = NULL,
  login_type = NULL,
  exchange_token = "access_token",
  provider = NULL,
  config = NULL
)
}
\arguments{
\item{host}{URL pointing to the openEO server service host}

\item{version}{the version number as string (optional)}

\item{user}{the user name (optional)}

\item{password}{the password (optional)}

\item{login_type}{either NULL, 'basic' or 'oidc'. This refers to the login mechanism that shall be used. NULL disables authentication.}

\item{exchange_token}{'access_token' or 'id_token' defines in the OIDC case the bearer token use}

\item{provider}{provider object as obtained by 'list_oidc_providers()'}

\item{config}{named list containing 'client_id' and 'sercret' or a path to the configuration file (type JSON)}
}
\description{
Connects to openEO service. If the backend provides a well-known endpoint that allows for redirecting to
specific versions, then you should provide the versions parameter.
}
\details{
Especially the \code{login_type} and the \code{authType} suggested by the client development guidelines are confusing. Here the login_type deals
just with considered login. Meaning 'basic' allows you to use username and password directly in the call, whereas 'oidc' will
open up a browser window, where you enter you credentials. The authentication against all protected endpoints will later
use the bearer token that the client has obtained after the login, unless the authentication was dropped with NULL anyways.

The parameter \code{version} is not required. If the service offers a well-known document of the
service the client will choose an appropriate version (default the most recent production ready version).

When calling this function the \code{\link{OpenEOClient}} is also stored in a variable in the package
which marks the latest service that was connected to.
}
\examples{
\dontrun{
# connect to a host with specific version and without authentication
con = connect(host='http://example.openeo.org',version='1.0.0-rc.2')

# connect to a host by direct url and basic login
con = connect(host='http://example.openeo.org/v1.0',
              user='user',
             password='password',
             login_type='basic')

# connect to a host with open id connect authentication
con = connect(host='http://example.openeo.org',
              version='1.0.0-rc.2',
              login_type='oidc')
}

}
\seealso{
\code{\link{active_connection}}
}
