% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{create_job}
\alias{create_job}
\title{Creates a job on the back-end from a Graph object}
\usage{
create_job(
  con = NULL,
  graph = NULL,
  title = NULL,
  description = NULL,
  plan = NULL,
  budget = NULL,
  ...
)
}
\arguments{
\item{con}{connected and authenticated openeo client (optional) otherwise \code{\link{active_connection}}
is used.}

\item{graph}{A Graph object}

\item{title}{Optional title of a job to be found}

\item{description}{Optional a more detailed information about a job}

\item{plan}{An optional execution plan offered by the back-end, determining how the job will be executed}

\item{budget}{An optional budget, which sets the maximum amount of credits to be used by the job}

\item{...}{additional configuration parameter for output generation}

\item{format}{The inteded format of the data to be returned}
}
\value{
the id of the job
}
\description{
This function shall be called after the user defined a process graph for the back-end to create a job on the
back-end. Therefore the user sends the process graph and the optional output specifications like
format and additional creation parameter by '...'. To add some meta data about the job, the user might use title or 
description. By providing a execution plan and a maximum usable budget the user can change the execution behavior of the
back-end provider.
}
