% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph.R
\name{create_process_graph}
\alias{create_process_graph}
\title{Stores a graph on the back-end}
\usage{
create_process_graph(con, graph, title = NULL, description = NULL)
}
\arguments{
\item{con}{connected and authorized openeo client object}

\item{graph}{a process graph definition}

\item{title}{the title of the process graph (optional)}

\item{description}{the description of a process graph (optional)}
}
\description{
Uploads the process graph information to the back-end and stores it for reuse.
}
