% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{create_service}
\alias{create_service}
\title{Prepares and publishes a service on the back-end}
\usage{
create_service(con = NULL, type, graph, title = NULL,
  description = NULL, enabled = NULL, parameters = NULL,
  plan = NULL, budget = NULL)
}
\arguments{
\item{con}{connected and authenticated openeo clien object (optional) otherwise \code{\link{active_connection}}
is used.}

\item{type}{character the ogc web service type name to be created}

\item{graph}{A process graph object}

\item{title}{character (optional) the title in human readabled form for the service}

\item{description}{character (optional) the description for the service}

\item{enabled}{logical}

\item{parameters}{a list of service creation parameters}

\item{plan}{character the billing plan}

\item{budget}{numeric the amount of credits that can be spent for this service}
}
\value{
service representation as list
}
\description{
The function will send a configuration object to the back-end to create a webservice from a job considering
additional parameter.
}
