% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_defined_processes.R
\name{create_user_process}
\alias{create_user_process}
\title{Stores a graph as user defined process at the back-end}
\usage{
create_user_process(
  graph,
  id,
  summary = NULL,
  description = NULL,
  submit = TRUE,
  con = NULL
)
}
\arguments{
\item{graph}{a process graph definition}

\item{id}{the title of the user process}

\item{summary}{the summary for the user process (optional)}

\item{description}{the description for the user process (optional)}

\item{submit}{whether to create a new user process at the openEO service or to create it for local use (default set to submit = TRUE)}

\item{con}{connected and authorized openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
a list assembling a process graph description or the graph id if send
}
\description{
Uploads the process graph information to the back-end and stores it for reuse as a user defined process.
}
