% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{delete_file}
\alias{delete_file}
\title{Delete a file from the user workspace}
\usage{
delete_file(src, con = NULL)
}
\arguments{
\item{src}{the relative filepath of the source file on the openEO back-end that shall be deleted}

\item{con}{authorized connection (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
logical
}
\description{
Sends a request to an openEO back-end in order to remove a specific file from the users workspaces.
}
