% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{delete_job}
\alias{delete_job}
\title{Delete a job}
\usage{
delete_job(con = NULL, job)
}
\arguments{
\item{con}{authenticated Connection (optional) otherwise \code{\link{active_connection}}
is used.}

\item{job}{the job or the id of the job}
}
\value{
logical with state of success
}
\description{
Deletes a job from the backend.
}
