% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{delete_job}
\alias{delete_job}
\title{Delete a job}
\usage{
delete_job(job, con = NULL)
}
\arguments{
\item{job}{the job or the id of the job}

\item{con}{authenticated Connection (optional) otherwise \code{\link[=active_connection]{active_connection()}}
is used.}
}
\value{
logical with state of success
}
\description{
Deletes a job from the back-end.
}
