% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph.R
\name{delete_process_graph}
\alias{delete_process_graph}
\title{Deletes a previously stored process graph}
\usage{
delete_process_graph(con, id)
}
\arguments{
\item{con}{connected and authorized openeo client object}

\item{id}{the id of the graph}
}
\description{
The function initiates the deletion of a process graph on the back-end. Only the owning user can delete
a graph. The graph also should not be part of any particular job.
}
