% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{delete_service}
\alias{delete_service}
\title{Deletes a service function for a job}
\usage{
delete_service(con = NULL, id)
}
\arguments{
\item{con}{connected and authorized openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}

\item{id}{the service id}
}
\description{
Queries the back-end and removes the current set service function of job.
}
