% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{deleteJob}
\alias{deleteJob}
\title{Delete a job}
\usage{
deleteJob(con, job_id)
}
\arguments{
\item{con}{authenticated Connection}

\item{job_id}{id of the job}
}
\value{
logical with state of success
}
\description{
Deletes a job from the backend.
}
