% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{deleteJob}
\alias{deleteJob}
\title{Deletes a job on the server}
\usage{
deleteJob(con, job_id)
}
\arguments{
\item{con}{An authenticated connection}

\item{job_id}{the id of the job on the server the user wants to connect to}
}
\value{
A success notification
}
\description{
Deletes a job on the server
}
