% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{deleteUserData}
\alias{deleteUserData}
\title{Deletes a file from the users workspace}
\usage{
deleteUserData(con, src)
}
\arguments{
\item{con}{authorized connection}

\item{src}{the relative filepath of the source file on the openeo backend that shall be deleted}
}
\value{
logical
}
\description{
Sends a request to an openeo backend in order to remove a specific file from the users workspaces
}
