% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{describe_collection}
\alias{describe_collection}
\title{Describe a collection}
\usage{
describe_collection(collection = NA, con = NULL)
}
\arguments{
\item{collection}{Collection object or the collections id}

\item{con}{Authentication object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
a Collection object with detailed information about a collection.
}
\description{
Queries an openEO back-end and retrieves a detailed description about one or more collections offered by the back-end.
}
