% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{describe_job}
\alias{describe_job}
\title{Fetches information about a job}
\usage{
describe_job(con = NULL, job)
}
\arguments{
\item{con}{authenticated Connection (optional) otherwise \code{\link{active_connection}}
is used.}

\item{job}{the job object or the id of the job}
}
\value{
a detailed description about the job
}
\description{
Returns a detailed description about a specified job. For example to check the status of a job.
}
