% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefined_processes.R
\name{describe_process}
\alias{describe_process}
\title{Describe a process}
\usage{
describe_process(process = NA, con = NULL)
}
\arguments{
\item{process}{id of a process to be described, the ProcessInfo object or a Process object}

\item{con}{Authentication object (optional) otherwise \code{\link[=active_connection]{active_connection()}}
is used.}
}
\value{
a list of detailed information
}
\description{
Queries an openEO back-end and retrieves more detailed information about offered processes
}
