% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph.R
\name{describe_process_graph}
\alias{describe_process_graph}
\title{Fetches the representation of a stored graph}
\usage{
describe_process_graph(con = NULL, id)
}
\arguments{
\item{con}{connected and authenticated openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}

\item{id}{The id of a process graph on the back-end}
}
\value{
the process graph as list
}
\description{
The function queries the back-end for a specific user defined process graph
}
