% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{describe_service}
\alias{describe_service}
\title{Describes a service}
\usage{
describe_service(con, id)
}
\arguments{
\item{con}{connected and authorized openeo client object}

\item{id}{the service id}
}
\value{
service as a list
}
\description{
Queries the server and returns information about a particular service
}
