% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{describe_service}
\alias{describe_service}
\title{Describes a service}
\usage{
describe_service(con = NULL, service)
}
\arguments{
\item{con}{connected and authorized openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}

\item{service}{the Service object or its id}
}
\value{
service as a list
}
\description{
Queries the server and returns information about a particular service
}
