% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{download_file}
\alias{download_file}
\title{Downloads a file from the users workspace}
\usage{
download_file(con, src, dst = NULL)
}
\arguments{
\item{con}{authorized connection}

\item{src}{the relative filepath of the source file on the openeo back-end}

\item{dst}{the destination file path on the local file system}
}
\value{
The file path of the stored file
}
\description{
Sends a request to an openeo back-end to access the users files and downloads them to a given location
}
