% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{download_results}
\alias{download_results}
\title{Downloads the results of a job into a specific folder}
\usage{
download_results(con = NULL, job, folder)
}
\arguments{
\item{con}{a connected and authenticated OpenEO connection (optional) otherwise \code{\link{active_connection}}
is used.}

\item{job}{job object or the job_id for which the results are fetched}

\item{folder}{a character string that is the target path on the local computer}
}
\value{
a list of the target file paths
}
\description{
The function will fetch the results of a asynchronous job and will download all files stated in the links. The parameter
'folder' will be the target location on the local computer.
}
