% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{executeTask}
\alias{executeTask}
\title{Executes a job directly and returns the data immediately}
\usage{
executeTask(con, task, format)
}
\arguments{
\item{con}{Connection}

\item{task}{A Process or chained processes to a Task}

\item{format}{The inteded format of the data to be returned}
}
\description{
Executes a job directly on the connected openEO backend and returns the data. It relates to
POST /api/jobs?evaluation="sync"
}
