% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{GeoJson}
\alias{GeoJson}
\title{GeoJson}
\value{
Object of \code{\link{R6Class}} representing an object in GeoJson.
}
\description{
Inheriting from \code{\link{Argument}} in order to represent a GeoJson object. This class represents geospatial features. 
Allowed values are either a list directly convertible into a valid GeoJson or polygon features from package 'sf'. 
If sf-objects are used, keep in mind that the objects are projected in  Lat/Long EPSG:4326, unless marked otherwise.
It is assumed that the coordinates have a projection. If the crs-object is set and does not match 
EPSG:4326, the polygon is transformed accordingly.
}
\seealso{
\code{\link{Array}}, \code{\link{Integer}}, \code{\link{EPSGCode}}, \code{\link{String}}, \code{\link{Number}}, 
\code{\link{Date}}, \code{\link{RasterCube}}, \code{\link{VectorCube}}, \code{\link{ProcessGraphArgument}}, 
\code{\link{ProcessGraphParameter}}, \code{\link{OutputFormatOptions}}, \code{\link{GeoJson}},
\code{\link{Boolean}}, \code{\link{DateTime}}, \code{\link{Time}}, \code{\link{BoundingBox}}, \code{\link{Kernel}}, 
\code{\link{TemporalInterval}}, \code{\link{TemporalIntervals}}, \code{\link{CollectionId}}, \code{\link{OutputFormat}},
\code{\link{AnyOf}}, \code{\link{ProjDefinition}}, \code{\link{UdfCodeArgument}}, \code{\link{UdfRuntimeArgument}} and 
\code{\link{UdfRuntimeVersionArgument}}
}
