% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{Kernel}
\alias{Kernel}
\title{Kernel}
\value{
Object of \code{\link{R6Class}} which represents a Kernel.
}
\description{
Inheriting from \code{\link{Argument}} in order to represent a 2 dimensional array of weights that shall be applied 
on the x and y (spatial) dimensions of the data cube. The inner level of the nested array aligns with the x axis and 
the outer level aligns with the y axis. Each level of the kernel must have an uneven number of elements.
}
\seealso{
\code{\link{Array}}, \code{\link{Integer}}, \code{\link{EPSGCode}}, \code{\link{String}}, \code{\link{Number}}, 
\code{\link{Date}}, \code{\link{RasterCube}}, \code{\link{VectorCube}}, \code{\link{ProcessGraphArgument}}, 
\code{\link{ProcessGraphParameter}}, \code{\link{OutputFormatOptions}}, \code{\link{GeoJson}},
\code{\link{Boolean}}, \code{\link{DateTime}}, \code{\link{Time}}, \code{\link{BoundingBox}}, \code{\link{Kernel}}, 
\code{\link{TemporalInterval}}, \code{\link{TemporalIntervals}}, \code{\link{CollectionId}}, \code{\link{OutputFormat}},
\code{\link{AnyOf}}, \code{\link{ProjDefinition}}, \code{\link{UdfCodeArgument}}, \code{\link{UdfRuntimeArgument}} and 
\code{\link{UdfRuntimeVersionArgument}}
}
