% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{Kernel}
\alias{Kernel}
\title{Kernel}
\value{
Object of \code{\link[R6:R6Class]{R6::R6Class()}} representing a Kernel.
}
\description{
Inheriting from \link{Argument} in order to represent a 2-dimensional array of weights applied
to the x and y (spatial) dimensions of the data cube. The inner level of the nested array is aligned to the x-axis and
the outer level is aligned to the y-axis. Each level of the kernel must have an uneven number of elements.
}
\seealso{
\link{Array}, \link{Integer}, \link{EPSGCode}, \link{String}, \link{Number},
\link{Date}, \link{DataCube}, \link{ProcessGraphArgument},
\link{ProcessGraphParameter}, \link{OutputFormatOptions}, \link{GeoJson},
\link{Boolean}, \link{DateTime}, \link{Time}, \link{BoundingBox}, \link{Kernel},
\link{TemporalInterval}, \link{TemporalIntervals}, \link{CollectionId}, \link{OutputFormat},
\link{AnyOf}, \link{ProjDefinition}, \link{UdfCodeArgument}, \link{UdfRuntimeArgument} and
\link{UdfRuntimeVersionArgument}, \link{MetadataFilter}
}
