% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{list_collections}
\alias{list_collections}
\title{List Data on conected server}
\usage{
list_collections(con = NULL)
}
\arguments{
\item{con}{Connection object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
object of class 'CollectionList'
}
\description{
List available collections stored on a openEO server and return them as a CollectionList which is a named list of Collection objects. 
The names are the collection IDs. Although the result at 'describe_collection' is also a Collection, this functions result will contain
lesser information than the detailed description received at the other function.
}
