% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{list_collections}
\alias{list_collections}
\title{List Data on conected server}
\usage{
list_collections(con)
}
\arguments{
\item{con}{Connection object}
}
\description{
List available collections stored on a openEO server
}
