% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_metadata.R
\name{list_features}
\alias{list_features}
\title{List the openEO endpoints}
\usage{
list_features(con = NULL)
}
\arguments{
\item{con}{A connected openEO client (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
data.frame or tibble (if available)
}
\description{
The client queries the version resolved back-end for its endpoint capabilities and returns it as
a tibble.
}
