% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_metadata.R
\name{list_features}
\alias{list_features}
\title{List the openeo endpoints}
\usage{
list_features(con)
}
\arguments{
\item{con}{A connected OpenEO client}
}
\value{
tibble
}
\description{
The client queries the version resolved back-end for its endpoint capabilities and returns it as
a tibble.
}
