% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_metadata.R
\name{list_file_formats}
\alias{list_file_formats}
\title{Returns the output formats}
\usage{
list_file_formats(con = NULL)
}
\arguments{
\item{con}{openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
a data frame with formats, the applied output data type ('raster', 'vector', 'table' and/or 'other') and optional configuration parameter
}
\description{
The function queries the back-end for supported output formats.
}
