% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{list_files}
\alias{list_files}
\title{Lists workspace files}
\usage{
list_files(con = NULL)
}
\arguments{
\item{con}{authorized connection (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
a tibble of for filenames and their sizes
}
\description{
Lists all files in the workspaces of the authenticated user.
}
