% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{list_jobs}
\alias{list_jobs}
\title{List the jobs that a user has}
\usage{
list_jobs(con = NULL)
}
\arguments{
\item{con}{the authenticated Connection (optional) otherwise \code{\link{active_connection}}
is used.}
}
\description{
lists the jobs that a user has uploaded or in execution
}
