% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_metadata.R
\name{list_service_types}
\alias{list_service_types}
\title{Returns the offered webservice types of the back-end}
\usage{
list_service_types(con = NULL)
}
\arguments{
\item{con}{a connected openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
a \code{ServiceTypeList}
}
\description{
The function queries the back-end for the supported webservice types that can be used on the client and returns a named list of
\code{ServiceType}, which are indexed by the service type ID. ServiceTypes can later be used when creating a supported web service
out of the user defined process (process graph).
}
