% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_metadata.R
\name{list_service_types}
\alias{list_service_types}
\title{Returns the web service types of the back-end}
\usage{
list_service_types(con = NULL)
}
\arguments{
\item{con}{a connected openEO client object (optional) otherwise \code{\link[=active_connection]{active_connection()}}
is used.}
}
\value{
a \code{ServiceTypeList}
}
\description{
The function queries the back-end for the supported web service types usable by the client and returns a named list of
\code{ServiceType} indexed by the service type ID. ServiceTypes can be used when creating a supported web service
from the user defined process (process graph).
}
