% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{list_services}
\alias{list_services}
\title{Lists the current users services}
\usage{
list_services(con = NULL)
}
\arguments{
\item{con}{connected and authenticated openEO client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
named list of Services (class ServiceList)
}
\description{
Queries the back-end to retrieve a list of services that the current user owns. Services are 
webservices like WCS, WFS, etc. The result is an object of type \code{ServiceList}, which is a named list of \code{Service}. The 
indices are the service IDs, the service object that is indexed by its ID and may use other functions instead of its service ID.
}
