% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf.R
\name{list_udf_runtimes}
\alias{list_udf_runtimes}
\title{Lists the supported UDF runtimes}
\usage{
list_udf_runtimes(con = NULL)
}
\arguments{
\item{con}{connected and authenticated openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
list of UDF runtimes with supported UDF types, versions and installed packages
}
\description{
The function queries the back-end for its supported UDF runtimes and returns detailed information about each
runtime.
}
