% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{listFiles}
\alias{listFiles}
\title{Lists workspace files}
\usage{
listFiles(con)
}
\arguments{
\item{con}{authorized connection}
}
\value{
a data.frame of for filenames and their size
}
\description{
Lists all files in the workspaces of the authenticated user.
}
