% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{log_job}
\alias{log_job}
\title{Job log}
\usage{
log_job(job, offset = NULL, limit = NULL, con = NULL)
}
\arguments{
\item{job}{the job or the job_id}

\item{offset}{the id of the log entry to start from}

\item{limit}{the limit of lines to be shown}

\item{con}{an optional connection if you want to address a specific service}
}
\value{
a \code{Log} object
}
\description{
Opens the log of job.
}
