% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{login}
\alias{login}
\title{Function to login to a specific backend}
\usage{
login(con, user = NULL, password = NULL)
}
\arguments{
\item{con}{connected back-end connection}

\item{user}{the user name}

\item{password}{the password}
}
\value{
a connected and authenticated back-end connection
}
\description{
Retrieves the bearer-token from the backend by sending user name and password to the backend. This step
is usually also performed in the "connect" step. But if you only connected to a back-end in order to 
register, then you need to log in afterwards.
}
