% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{MetadataFilter}
\alias{MetadataFilter}
\title{MetadataFilter}
\value{
Object of \code{\link{R6Class}} representing a list of \code{\link{Process}} in order to filter for collections.
}
\description{
Inheriting from \code{\link{ProcessGraphArgument}} in order to represent a list of functions that is internally 
interpreted into \code{\link{Process}} objects.
}
\examples{
\dontrun{
# define filter statement
filter = list(
   "eo:cloud_cover" = function(x) x >= 0 & x < 50, 
   "platform" = function(x) x == "Sentinel-2A"
)

# setting the arguments is done via the process graph building with of 'processes()'
}

}
\seealso{
\code{\link{Array}}, \code{\link{Integer}}, \code{\link{EPSGCode}}, \code{\link{String}}, \code{\link{Number}}, 
\code{\link{Date}}, \code{\link{RasterCube}}, \code{\link{VectorCube}}, \code{\link{ProcessGraphArgument}}, 
\code{\link{ProcessGraphParameter}}, \code{\link{OutputFormatOptions}}, \code{\link{GeoJson}},
\code{\link{Boolean}}, \code{\link{DateTime}}, \code{\link{Time}}, \code{\link{BoundingBox}}, \code{\link{Kernel}}, 
\code{\link{TemporalInterval}}, \code{\link{TemporalIntervals}}, \code{\link{CollectionId}}, \code{\link{OutputFormat}},
\code{\link{AnyOf}}, \code{\link{ProjDefinition}}, \code{\link{UdfCodeArgument}}, \code{\link{UdfRuntimeArgument}} and 
\code{\link{UdfRuntimeVersionArgument}},\code{\link{TemporalIntervals}}, \code{\link{MetadataFilter}}
}
