% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{OIDCAuth}
\alias{OIDCAuth}
\title{OIDC Authentication}
\description{
A class that authenticates via Open ID Connect. It inherits all fields and functions from \code{\link{IAuth}}. The OIDC login
will open a browser window where you will be asked to enter your credentials. The website belongs to the OIDC provider of the
chosen openeo back-end. Meanwhile the client will start a server demon in the background that listens for the callback from
the OIDC provider (the R client has provided its ID and the callback URI beforehand).
The \code{access_token} will be returned when queried. If the lease time has run out the client will refresh the access_token
automatically.
}
\section{Methods}{

\describe{
  \item{\code{$new(host,discovery_document = NULL)}}{the constructor for the authentication}
  \item{\code{$getUserData()}}{queries the OIDC provider for the user data like the 'user_id'}
}
}

\section{Arguments}{

\describe{
  \item{\code{host}}{the URL pointing to the OIDC discovery document}
  \item{\code{discover_document}}{optional the discovery document of the OIDC endpoint}
}
}

