% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{openeo-deprecated}
\alias{openeo-deprecated}
\title{openeo-deprecated}
\description{
Lists all currently deprecated functions that will be removed in the future.
}
\section{Deprecated}{

\describe{
\item{graphToJSON(x,...)}{replaced by toJSON}
\item{processToJSON(x,...)}{replaced by toJSON}
}
}

