% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{orderResult}
\alias{orderResult}
\title{Defines a job and starts the evaluation}
\usage{
orderResult(con, task = NULL, graph_id = NULL, format, ...)
}
\arguments{
\item{con}{connected and authenticated openeo client}

\item{task}{A Process or chained processes to a Task}

\item{graph_id}{The id of an already stored process graph on the same back-end}

\item{format}{The inteded format of the data to be returned}

\item{...}{additional configuration parameter for output generation}
}
\value{
the job_id of the defined job
}
\description{
Function to execute two operations in one call. First define a job on the server, then call
the endpoint to execute it immediately on the back-end.
}
