% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{OutputFormat}
\alias{OutputFormat}
\title{OutputFormat class}
\value{
Object of \code{\link{R6Class}} which represents an output format of a back-end.
}
\description{
Inheriting from \code{\link{Argument}} in order to represent an output format of a back-end as  a 
character string value.
}
\seealso{
\code{\link{Array}}, \code{\link{Integer}}, \code{\link{EPSGCode}}, \code{\link{String}}, \code{\link{Number}}, 
\code{\link{Date}}, \code{\link{RasterCube}}, \code{\link{VectorCube}}, \code{\link{Callback}}, 
\code{\link{CallbackValue}}, \code{\link{Variable}}, \code{\link{OutputFormatOptions}}, \code{\link{GeoJson}},
\code{\link{Boolean}}, \code{\link{DateTime}}, \code{\link{Time}}, \code{\link{BoundingBox}}, \code{\link{Kernel}}, 
\code{\link{TemporalInterval}}, \code{\link{temporalIntervals}}, \code{\link{CollectionId}}, \code{\link{OutputFormat}},
\code{\link{AnyOf}} and \code{\link{ProjDefinition}}
}
