% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph_building.R
\name{parse_graph}
\alias{parse_graph}
\title{Converts a JSON openEO graph into an R graph}
\usage{
parse_graph(json, parameters = NULL, con = NULL)
}
\arguments{
\item{json}{the json graph in a textual representation or an already parsed list object}

\item{parameters}{optional parameters}

\item{con}{a connected openEO client (optional) otherwise \code{\link[=active_connection]{active_connection()}}
is used.}
}
\value{
Graph object
}
\description{
The function reads and parses a json text and creates a Graph object.
}
