% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph_building.R
\name{parse_graph}
\alias{parse_graph}
\title{Parses a JSON openeo graph into a R graph}
\usage{
parse_graph(json, parameters = NULL, con = NULL)
}
\arguments{
\item{json}{the json graph in a textual representation or an already parsed list object}

\item{parameters}{optional parameters}

\item{con}{a connected openeo client (optional) otherwise \code{\link{active_connection}}
is used.}
}
\value{
Graph object
}
\description{
The function reads and parses the given json text and creates based on the information of the
text a Graph object.
}
