% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph_building.R
\name{Process}
\alias{Process}
\title{Process object}
\value{
Object of \code{\link{R6Class}} with methods for storing meta data of back-end processes and user assigned data
}
\description{
This object reflects the process offered by a back-end. It will be created with the information of a received 
JSON object for a single process, after the arguments of the process have been translated into Argument objects.
}
\section{Fields}{

\describe{
\item{\code{parameters}}{a named list of Argument objects}
}}

\section{Methods}{

\describe{
   \item{$new(id,parameters,description=character(), summary = character(), parameter_order=character(),returns)}{}
   \item{$getId()}{returns the id of a process which was defined on the back-end}
   \item{$getParameters()}{returns a named list of Arguments}
   \item{$getReturns()}{returns the schema for the return type as list}
   \item{$getFormals()}{returns the function formals for this process - usually a name vector of NAs where the name 
   corresponds to the parameter name}
   \item{$validate()}{validates the processes argument values}
   \item{$serialize()}{serializes the process - mainly used as primary serialization for a \code{\link{ProcessNode}}}
   \item{$setParameter(name,value)}{sets the value of a parameter}
   \item{$getParameter(name)}{returns the Argument object with the provided name}
   \item{$setDescription(value)}{sets the description text}
   \item{$getCharacteristics()}{select all non functions of the private area, to be used when copying process 
   information into a process node}
}
}

\section{Arguments}{

\describe{
   \item{id}{process id from the back-end}
   \item{parameters}{a list of Argument objects}
   \item{description}{the process description}
   \item{summary}{the summary of a process}
   \item{parameter_order}{the order in which the parameters have to be returned to the backend}
   \item{returns}{the schema part of the result definition}
   \item{name}{a parameter name}
   \item{value}{the value for a parameter or the description text}
}
}

