% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph_building.R
\name{process}
\alias{process}
\title{Create a process object}
\usage{
process(process = NULL, process_id, prior.name = "imagery", ...)
}
\arguments{
\item{process}{A process that might be chained with this process}

\item{process_id}{ID of the process offered by the connected openEO backend}

\item{prior.name}{The name of the prior process / collection, default "imagery"}

\item{...}{named arguments that are passed to the process description}
}
\description{
defines a process with arguments
}
