% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph.R
\name{process_graph_builder}
\alias{process_graph_builder}
\title{Get a process graph builder from the connection}
\usage{
process_graph_builder(con)
}
\arguments{
\item{con}{a connection to an openeo back-end}
}
\value{
a Graph object with the offered processes of the backend
}
\description{
Queries the connected back-end for all available processes and collection names and registers them via R functions on
a Graph object to model a process graph in R.
}
