% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph_building.R
\name{ProcessCollection}
\alias{ProcessCollection}
\title{Process Collection}
\description{
This object contains template functions for process graph building from the processes offered by an openEO service
}
\section{Methods}{

\describe{
   \item{\code{$new(con = NULL)}}{The object creator created an openEO connection.} 
}
}

\section{Arguments}{

\describe{
   \item{con}{optional an active and authenticated Connection (optional) otherwise \code{\link{active_connection}}
is used.}
}
}

