% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefined_processes.R
\name{ProcessCollection}
\alias{ProcessCollection}
\title{Process Collection}
\description{
This object contains template functions for process graph building from the processes offered by an openEO service. This object is
an unlocked R6 object, in order to add new functions at runtime.
}
\section{Methods}{

\describe{
\item{\verb{$new(con = NULL)}}{The object creator created an openEO connection.}
}
}

\section{Arguments}{

\describe{
\item{con}{optional an active and authenticated Connection (optional) otherwise \code{\link[=active_connection]{active_connection()}}
is used.}
}
}

\seealso{
\code{\link[=processes]{processes()}}
}
