% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph_building.R
\name{ProcessCollection}
\alias{ProcessCollection}
\title{Process Collection}
\description{
Similarly to the Graph object this object contains template functions for process graph building, 
but without the overload of creating a Graph object, which contains ProcessNodes.
}
\section{Fields}{

\describe{
\item{\code{data}}{a named list of collection ids or process graph parameters depending on the context}
}}

\section{Methods}{

\describe{
   \item{\code{$new(con = NULL, data = list())}}{The object creator created from processes and available data. 
   If \code{data} was omitted then it is fetched from \code{\link{list_collections}}. }
}
}

\section{Arguments}{

\describe{
   \item{data}{optional a named list of available data}
}
}

