% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{RasterCube}
\alias{RasterCube}
\title{RasterCube}
\value{
Object of \code{\link{R6Class}} which represents a raster cube.
}
\description{
Inheriting from \code{\link{Argument}} in order to represent a raster cube. This is usually the in- and 
output format of a process. Unless the process operates within a ProcessGraph on reduced data. Analogous to
this the \code{\link{VectorCube}} behaves in the same manner, but with spatial feature data.
}
\seealso{
\code{\link{Array}}, \code{\link{Integer}}, \code{\link{EPSGCode}}, \code{\link{String}}, \code{\link{Number}}, 
\code{\link{Date}}, \code{\link{RasterCube}}, \code{\link{VectorCube}}, \code{\link{ProcessGraph}}, 
\code{\link{ProcessGraphParameter}}, \code{\link{OutputFormatOptions}}, \code{\link{GeoJson}},
\code{\link{Boolean}}, \code{\link{DateTime}}, \code{\link{Time}}, \code{\link{BoundingBox}}, \code{\link{Kernel}}, 
\code{\link{TemporalInterval}}, \code{\link{TemporalIntervals}}, \code{\link{CollectionId}}, \code{\link{OutputFormat}},
\code{\link{AnyOf}} and \code{\link{ProjDefinition}}
}
