% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{start_job}
\alias{start_job}
\title{Starts remote asynchronous evaluation of a job}
\usage{
start_job(con, job)
}
\arguments{
\item{con}{connected and authenticated openeo client}

\item{job}{the job object or the job id of the defined job}
}
\value{
the job_id of the defined job
}
\description{
The function sends a start signal to the backend in order to start processing the results
for a defined job.
}
