% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities-mapping.R
\name{supports}
\alias{supports}
\title{Tag support lookup}
\usage{
supports(con = NULL, tag_name)
}
\arguments{
\item{con}{backend connection (optional) otherwise \code{\link[=active_connection]{active_connection()}}
is used.}

\item{tag_name}{the endpoints 'tag' name as character}
}
\value{
logical - whether the back-end supports the endpoint or not
}
\description{
Finds the client tag for a particular endpoint on the back-end and returns whether it is available
or not.
}
