% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{TemporalIntervals}
\alias{TemporalIntervals}
\title{TemporalIntervals}
\value{
Object of \code{\link[R6:R6Class]{R6::R6Class()}} representing a list of temporal intervals.
}
\description{
Inheriting from \link{Argument} in order to represent a list of \link{TemporalInterval}.
}
\seealso{
\link{Array}, \link{Integer}, \link{EPSGCode}, \link{String}, \link{Number},
\link{Date}, \link{DataCube}, \link{ProcessGraphArgument},
\link{ProcessGraphParameter}, \link{OutputFormatOptions}, \link{GeoJson},
\link{Boolean}, \link{DateTime}, \link{Time}, \link{BoundingBox}, \link{Kernel},
\link{TemporalInterval}, \link{TemporalIntervals}, \link{CollectionId}, \link{OutputFormat},
\link{AnyOf}, \link{ProjDefinition}, \link{UdfCodeArgument}, \link{UdfRuntimeArgument} and
\link{UdfRuntimeVersionArgument}, \link{MetadataFilter}
}
