% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_types.R
\name{UdfCodeArgument}
\alias{UdfCodeArgument}
\title{UdfCodeArgument class}
\value{
Object of \code{\link{R6Class}} is an argument that expects an UDF code or a file path.
}
\description{
Inheriting from \code{\link{Argument}} in order to represent the UDF code that will be executed in a UDF call. The script has to 
be passed as a character string or as a local file path from which the script can be loaded.
}
\seealso{
\code{\link{Array}}, \code{\link{Integer}}, \code{\link{EPSGCode}}, \code{\link{String}}, \code{\link{Number}}, 
\code{\link{Date}}, \code{\link{RasterCube}}, \code{\link{VectorCube}}, \code{\link{ProcessGraphArgument}}, 
\code{\link{ProcessGraphParameter}}, \code{\link{OutputFormatOptions}}, \code{\link{GeoJson}},
\code{\link{Boolean}}, \code{\link{DateTime}}, \code{\link{Time}}, \code{\link{BoundingBox}}, \code{\link{Kernel}}, 
\code{\link{TemporalInterval}}, \code{\link{TemporalIntervals}}, \code{\link{CollectionId}}, \code{\link{OutputFormat}},
\code{\link{AnyOf}}, \code{\link{ProjDefinition}}, \code{\link{UdfCodeArgument}}, \code{\link{UdfRuntimeArgument}} and 
\code{\link{UdfRuntimeVersionArgument}},\code{\link{TemporalIntervals}}, \code{\link{MetadataFilter}}
}
