% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{update_job}
\alias{update_job}
\title{Modifies a job with given parameter}
\usage{
update_job(con = NULL, id, title = NULL, description = NULL,
  process_graph = NULL, plan = NULL, budget = NULL, format = NULL,
  ...)
}
\arguments{
\item{con}{connected and authenticated openeo client (optional) otherwise \code{\link{active_connection}}
is used.}

\item{id}{the job id of a created job}

\item{title}{update title for the job}

\item{description}{update description}

\item{process_graph}{a Graph object created with the process_graph_builder}

\item{plan}{replaces plan with the set value}

\item{budget}{replaces or sets the credits that can be spent at maximum}

\item{format}{the output format}

\item{...}{The create options parameter you want to change. See Details for more information}
}
\description{
The function will mofidy a stored job with the given parameter. The dot parameter will contain all the values
that shall be replaced or removed. Shows a message of result or failure.
}
\details{
The '...' operator shall contain all the values that are to be replaced in the job. There are some reserved
keys. 
'process_graph' will replace the process graph with a newly defined one, therefore the process graph needs to be a Graph object.
'format' will change the desired output format.
All other parameter will be assumed to be special output parameter. Remember, you don't need to specify a process graph or graph_id,
e.g. if you just want to update the output format. 
To leave parameter unchanged, then don't mention it in the ... parameter. If you want to delete some, then set them to NULL.
}
