% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_defined_processes.R
\name{update_user_process}
\alias{update_user_process}
\title{Update the current process graph}
\usage{
update_user_process(
  id,
  graph = NULL,
  summary = NULL,
  description = NULL,
  con = NULL,
  ...
)
}
\arguments{
\item{id}{process graph id}

\item{graph}{a process graph definition created by combining 'process()', 'collection()' or using a ProcessGraphBuilder}

\item{summary}{summary of the process graph (optional)}

\item{description}{description of the process graph (optional)}

\item{con}{connected and authorized openeo client object (optional) otherwise \code{\link{active_connection}}
is used.}

\item{...}{additional parameters passed to toJSON (like 'digits')}
}
\description{
Upload a process graph to the back-end under a given (existing) process graph.
}
