% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_defined_processes.R
\name{user_processes}
\alias{user_processes}
\title{Process collection for user defined processes}
\usage{
user_processes(con = NULL)
}
\arguments{
\item{con}{a connection to an openeo back-end (optional) otherwise \code{\link{active_connection}}
is used in order to access personal user defined processes you need to be logged in}
}
\value{
\code{\link{UserProcessCollection}}
}
\description{
The created process graphs via \code{\link{create_process_graph}} at the openEO service are user defined processes. 
This means that they can be used within the creation of process graphs themselves. For processes provided by the 
particular openEO service the \code{\link{processes}} function can be used to obtain a builder for those processes. 
Analoguous to this idea, this function creates a builder object for user defined proceses which are listed and described
with \code{\link{get_user_process}} and \code{\link{list_user_processes}}.
}
