% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph_building.R
\name{UserProcessCollection}
\alias{UserProcessCollection}
\title{User Defined Process Collection}
\description{
This object contains template functions for process graph building from the user defined processes at an openEO service. 
This object is an R6 object that is not locked, in order to add new functions at runtime, in the same fashion as 
\code{\link{ProcessCollection}} for predefined processes.
}
\section{Methods}{

\describe{
   \item{\code{$new(con = NULL)}}{The object creator created an openEO connection.} 
}
}

\section{Arguments}{

\describe{
   \item{con}{optional an active and authenticated Connection (optional) otherwise \code{\link{active_connection}}
is used.}
}
}

