% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph_building.R
\name{variables}
\alias{variables}
\title{Lists the defined variables for a graph}
\usage{
variables(x)
}
\arguments{
\item{x}{a process graph object or a process node}
}
\value{
a named list of Variables
}
\description{
The function creates a list of the defined (but not necessarily used) variables of a process graph.
}
